+++ db.py
@@ -1,1804 +1,58 @@
 import sqlite3
 import uuid
-import hashlib
+# hashlib has been removed as it's no longer used in this file
 from datetime import datetime
 import json
 import os # Added os import
 
-# Global variable for the database name
-DATABASE_NAME = "app_data.db"
-
-# Constants for document context paths
-# Assuming db.py is in a subdirectory of the app root (e.g., /app/core/db.py or /app/db.py)
-# If db.py is directly in /app, then os.path.dirname(__file__) is /app
-# If db.py is in /app/core, then os.path.dirname(__file__) is /app/core, so "../" goes to /app
-APP_ROOT_DIR_CONTEXT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-LOGO_SUBDIR_CONTEXT = "company_logos" # Example, adjust if actual path is different, e.g., "static/logos" or "data/logos"
-
-
-def initialize_database():
-    """
-    Initializes the database by creating tables if they don't already exist.
-    """
-    conn = sqlite3.connect(DATABASE_NAME)
-    conn.row_factory = sqlite3.Row # <-- ADD THIS LINE
-    cursor = conn.cursor()
-
-    # Create Users table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Users (
-        user_id TEXT PRIMARY KEY, 
-        username TEXT NOT NULL UNIQUE,
-        password_hash TEXT NOT NULL,
-        full_name TEXT,
-        email TEXT NOT NULL UNIQUE,
-        role TEXT NOT NULL, -- e.g., 'admin', 'manager', 'member'
-        is_active BOOLEAN DEFAULT TRUE,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        last_login_at TIMESTAMP
-    )
-    """)
-
-    # Create Companies table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Companies (
-        company_id TEXT PRIMARY KEY,
-        company_name TEXT NOT NULL,
-        address TEXT,
-        payment_info TEXT,
-        logo_path TEXT,
-        other_info TEXT,
-        is_default BOOLEAN DEFAULT FALSE,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-    )
-    """)
-
-    # Create CompanyPersonnel table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS CompanyPersonnel (
-        personnel_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        company_id TEXT NOT NULL,
-        name TEXT NOT NULL,
-        role TEXT NOT NULL, -- e.g., "seller", "technical_manager"
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (company_id) REFERENCES Companies (company_id) ON DELETE CASCADE
-    )
-    """)
-
-    # Create TeamMembers table (New, depends on Users)
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS TeamMembers (
-        team_member_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        user_id TEXT UNIQUE,      -- Link to Users table, can be NULL
-        full_name TEXT NOT NULL,
-        email TEXT NOT NULL UNIQUE,
-        role_or_title TEXT,       -- e.g., 'Project Manager', 'Developer', 'Designer'
-        department TEXT,          -- e.g., 'Engineering', 'Design', 'Sales'
-        phone_number TEXT,
-        profile_picture_url TEXT,
-        is_active BOOLEAN DEFAULT TRUE,
-        notes TEXT,
-        hire_date TEXT,
-        performance INTEGER DEFAULT 0,
-        skills TEXT,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (user_id) REFERENCES Users (user_id) ON DELETE SET NULL -- If user is deleted, set user_id to NULL
-    )
-    """)
-
-    # Create Countries table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Countries (
-        country_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        country_name TEXT NOT NULL UNIQUE
-    )
-    """)
-
-    # Create Cities table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Cities (
-        city_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        country_id INTEGER NOT NULL,
-        city_name TEXT NOT NULL,
-        FOREIGN KEY (country_id) REFERENCES Countries (country_id)
-    )
-    """)
-
-    # --- StatusSettings Table ---
-    # Check if StatusSettings table exists and if icon_name column is present
-    cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='StatusSettings'")
-    table_exists = cursor.fetchone()
-
-    if table_exists:
-        cursor.execute("PRAGMA table_info(StatusSettings)")
-        columns = [column[1] for column in cursor.fetchall()]
-        if 'icon_name' not in columns:
-            print("StatusSettings table exists but icon_name column is missing. Adding it now.")
-            cursor.execute("ALTER TABLE StatusSettings ADD COLUMN icon_name TEXT")
-            conn.commit() # Commit the schema change immediately
-        else:
-            print("StatusSettings table and icon_name column already exist.")
-    else:
-        print("StatusSettings table does not exist. It will be created.")
-
-    # Create StatusSettings table (IF NOT EXISTS handles the case where it was just created or existed)
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS StatusSettings (
-        status_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        status_name TEXT NOT NULL,
-        status_type TEXT NOT NULL, -- e.g., 'Client', 'Project', 'Task'
-        color_hex TEXT,
-        icon_name TEXT, -- Ensured by above logic or created here
-        default_duration_days INTEGER,
-        is_archival_status BOOLEAN DEFAULT FALSE,
-        is_completion_status BOOLEAN DEFAULT FALSE,
-        UNIQUE (status_name, status_type)
-    )
-    """)
-
-    # --- Pre-populate StatusSettings ---
-    # Values include icon_name, ensuring it's populated.
-    default_statuses = [
-        # Client Statuses
-        {'status_name': 'En cours', 'status_type': 'Client', 'color_hex': '#3498db', 'icon_name': 'dialog-information', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'Prospect', 'status_type': 'Client', 'color_hex': '#f1c40f', 'icon_name': 'user-status-pending', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'Actif', 'status_type': 'Client', 'color_hex': '#2ecc71', 'icon_name': 'user-available', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'Inactif', 'status_type': 'Client', 'color_hex': '#95a5a6', 'icon_name': 'user-offline', 'is_completion_status': False, 'is_archival_status': True},
-        {'status_name': 'Complété', 'status_type': 'Client', 'color_hex': '#27ae60', 'icon_name': 'task-complete', 'is_completion_status': True, 'is_archival_status': False},
-        {'status_name': 'Archivé', 'status_type': 'Client', 'color_hex': '#7f8c8d', 'icon_name': 'archive', 'is_completion_status': False, 'is_archival_status': True},
-        {'status_name': 'Urgent', 'status_type': 'Client', 'color_hex': '#e74c3c', 'icon_name': 'dialog-warning', 'is_completion_status': False, 'is_archival_status': False},
-
-        # Project Statuses
-        {'status_name': 'Planning', 'status_type': 'Project', 'color_hex': '#1abc9c', 'icon_name': 'view-list-bullet', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'En cours', 'status_type': 'Project', 'color_hex': '#3498db', 'icon_name': 'view-list-ordered', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'En attente', 'status_type': 'Project', 'color_hex': '#f39c12', 'icon_name': 'view-list-remove', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'Terminé', 'status_type': 'Project', 'color_hex': '#2ecc71', 'icon_name': 'task-complete', 'is_completion_status': True, 'is_archival_status': False},
-        {'status_name': 'Annulé', 'status_type': 'Project', 'color_hex': '#c0392b', 'icon_name': 'dialog-cancel', 'is_completion_status': False, 'is_archival_status': True},
-        {'status_name': 'En pause', 'status_type': 'Project', 'color_hex': '#8e44ad', 'icon_name': 'media-playback-pause', 'is_completion_status': False, 'is_archival_status': False},
-
-        # Task Statuses
-        {'status_name': 'To Do', 'status_type': 'Task', 'color_hex': '#bdc3c7', 'icon_name': 'view-list-todo', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'In Progress', 'status_type': 'Task', 'color_hex': '#3498db', 'icon_name': 'view-list-progress', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'Done', 'status_type': 'Task', 'color_hex': '#2ecc71', 'icon_name': 'task-complete', 'is_completion_status': True, 'is_archival_status': False},
-        {'status_name': 'Blocked', 'status_type': 'Task', 'color_hex': '#e74c3c', 'icon_name': 'dialog-error', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'Review', 'status_type': 'Task', 'color_hex': '#f1c40f', 'icon_name': 'view-list-search', 'is_completion_status': False, 'is_archival_status': False},
-        {'status_name': 'Cancelled', 'status_type': 'Task', 'color_hex': '#7f8c8d', 'icon_name': 'dialog-cancel', 'is_completion_status': False, 'is_archival_status': True}
-    ]
-
-    for status in default_statuses:
-        # Using INSERT OR REPLACE to ensure existing statuses are updated with latest values (e.g., icon_name, color_hex)
-        # The UNIQUE constraint (status_name, status_type) is used by REPLACE.
-        cursor.execute("""
-            INSERT OR REPLACE INTO StatusSettings (
-                status_name, status_type, color_hex, icon_name,
-                is_completion_status, is_archival_status, default_duration_days -- Added default_duration_days if applicable
-            ) VALUES (?, ?, ?, ?, ?, ?, ?)
-        """, (
-            status['status_name'],
-            status['status_type'],
-            status['color_hex'],
-            status.get('icon_name'),
-            status.get('is_completion_status', False),
-            status.get('is_archival_status', False),
-            status.get('default_duration_days') # Ensure this key exists in your default_statuses or handle if missing
-        ))
-    # Note: If default_duration_days is not always present in default_statuses,
-    # you might need to adjust the INSERT OR REPLACE or ensure the data structure is consistent.
-    # For now, assuming it might be there or defaults to NULL if not provided and column allows NULL.
-
-    # Create Clients tabhhhle
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Clients (
-        client_id TEXT PRIMARY KEY,
-        client_name TEXT NOT NULL,
-        company_name TEXT,
-        primary_need_description TEXT, -- Maps to 'need' from main.py
-        project_identifier TEXT NOT NULL, -- Added from main.py
-        country_id INTEGER,
-        city_id INTEGER,
-        default_base_folder_path TEXT UNIQUE, -- Added UNIQUE from main.py's base_folder_path
-        status_id INTEGER,
-        selected_languages TEXT, -- Comma-separated list of language codes
-        price REAL DEFAULT 0, -- Added from main.py
-        notes TEXT,
-        category TEXT,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP, -- Maps to creation_date from main.py
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP, -- Maps to last_modified from main.py
-        created_by_user_id TEXT,
-        FOREIGN KEY (country_id) REFERENCES Countries (country_id),
-        FOREIGN KEY (city_id) REFERENCES Cities (city_id),
-        FOREIGN KEY (status_id) REFERENCES StatusSettings (status_id),
-        FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id)
-    )
-    """)
-
-    # Create ClientNotes table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ClientNotes (
-        note_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        client_id TEXT NOT NULL,
-        timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        note_text TEXT NOT NULL,
-        user_id TEXT,
-        FOREIGN KEY (client_id) REFERENCES Clients (client_id) ON DELETE CASCADE,
-        FOREIGN KEY (user_id) REFERENCES Users (user_id) ON DELETE SET NULL
-    )
-    """)
-
-    # Create Projects table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Projects (
-        project_id TEXT PRIMARY KEY,
-        client_id TEXT NOT NULL,
-        project_name TEXT NOT NULL,
-        description TEXT,
-        start_date DATE,
-        deadline_date DATE,
-        budget REAL,
-        status_id INTEGER,
-        progress_percentage INTEGER DEFAULT 0,
-        manager_team_member_id TEXT, -- Assuming this refers to a User ID
-        priority INTEGER DEFAULT 0,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (client_id) REFERENCES Clients (client_id),
-        FOREIGN KEY (status_id) REFERENCES StatusSettings (status_id),
-        FOREIGN KEY (manager_team_member_id) REFERENCES Users (user_id)
-    )
-    """)
-
-    # Create Contacts table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Contacts (
-        contact_id INTEGER PRIMARY KEY AUTOINCREMENT, -- Changed to INTEGER for AUTOINCREMENT
-        name TEXT NOT NULL, -- This will be used as a fallback if displayName is not available
-        email TEXT UNIQUE,
-        phone TEXT, -- This will be the primary phone, specific type in phone_type
-        position TEXT, -- Kept for general position, specific title in organization_title
-        company_name TEXT, -- Kept for general company name, specific in organization_name
-        notes TEXT, -- Used for notes_text
-        givenName TEXT,
-        familyName TEXT,
-        displayName TEXT,
-        phone_type TEXT, -- e.g., 'work', 'mobile', 'home'
-        email_type TEXT, -- e.g., 'work', 'personal'
-        address_formattedValue TEXT,
-        address_streetAddress TEXT,
-        address_city TEXT,
-        address_region TEXT,
-        address_postalCode TEXT,
-        address_country TEXT,
-        organization_name TEXT,
-        organization_title TEXT,
-        birthday_date TEXT, -- Store as TEXT in ISO format e.g., YYYY-MM-DD or MM-DD
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-    )
-    """)
-
-    # Create ClientContacts table (associative table)
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ClientContacts (
-        client_contact_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        client_id TEXT NOT NULL,
-        contact_id TEXT NOT NULL,
-        is_primary_for_client BOOLEAN DEFAULT FALSE,
-        can_receive_documents BOOLEAN DEFAULT TRUE,
-        FOREIGN KEY (client_id) REFERENCES Clients (client_id) ON DELETE CASCADE, -- Added ON DELETE CASCADE
-        FOREIGN KEY (contact_id) REFERENCES Contacts (contact_id) ON DELETE CASCADE, -- Added ON DELETE CASCADE
-        UNIQUE (client_id, contact_id)
-    )
-    """)
-
-    # Create Products table (New)
-    # Idempotent ALTER TABLE for existing Products table
-    try:
-        cursor.execute("PRAGMA table_info(Products)")
-        columns_info = cursor.fetchall()
-        existing_column_names = {info['name'] for info in columns_info} # Assuming conn.row_factory = sqlite3.Row
-
-        altered = False
-        if 'weight' not in existing_column_names:
-            cursor.execute("ALTER TABLE Products ADD COLUMN weight REAL")
-            print("Added 'weight' column to Products table.")
-            altered = True
-        if 'dimensions' not in existing_column_names:
-            cursor.execute("ALTER TABLE Products ADD COLUMN dimensions TEXT")
-            print("Added 'dimensions' column to Products table.")
-            altered = True
-
-        if altered:
-            conn.commit() # Commit ALTER TABLE statements immediately
-
-    except sqlite3.Error as e:
-        # This might happen if Products table doesn't exist yet, which is fine.
-        # The CREATE TABLE IF NOT EXISTS below will handle it.
-        print(f"Info: Checking columns for Products table, it might not exist yet or other error: {e}")
-
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Products (
-        product_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        product_name TEXT NOT NULL,
-        description TEXT,
-        category TEXT, 
-        language_code TEXT DEFAULT 'fr',
-        base_unit_price REAL NOT NULL,
-        unit_of_measure TEXT, 
-        weight REAL, -- New column
-        dimensions TEXT, -- New column (e.g., "LxWxH cm")
-        is_active BOOLEAN DEFAULT TRUE,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        UNIQUE (product_name, language_code)
-    )
-    """)
-
-    # Create ProductEquivalencies table (New)
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ProductEquivalencies (
-        equivalence_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        product_id_a INTEGER NOT NULL,
-        product_id_b INTEGER NOT NULL,
-        FOREIGN KEY (product_id_a) REFERENCES Products (product_id) ON DELETE CASCADE,
-        FOREIGN KEY (product_id_b) REFERENCES Products (product_id) ON DELETE CASCADE,
-        UNIQUE (product_id_a, product_id_b)
-    )
-    """)
-
-    # Create ClientProjectProducts table (New - Association)
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ClientProjectProducts (
-        client_project_product_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        client_id TEXT NOT NULL,
-        project_id TEXT, 
-        product_id INTEGER NOT NULL,
-        quantity INTEGER NOT NULL DEFAULT 1,
-        unit_price_override REAL, 
-        total_price_calculated REAL, 
-        added_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (client_id) REFERENCES Clients (client_id) ON DELETE CASCADE,
-        FOREIGN KEY (project_id) REFERENCES Projects (project_id) ON DELETE CASCADE,
-        FOREIGN KEY (product_id) REFERENCES Products (product_id) ON DELETE CASCADE,
-        UNIQUE (client_id, project_id, product_id)
-    )
-    """)
-
-    # Create ScheduledEmails table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ScheduledEmails (
-        scheduled_email_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        recipient_email TEXT NOT NULL,
-        subject TEXT NOT NULL,
-        body_html TEXT,
-        body_text TEXT,
-        scheduled_send_at TIMESTAMP NOT NULL, 
-        status TEXT NOT NULL DEFAULT 'pending', 
-        sent_at TIMESTAMP, 
-        error_message TEXT,
-        related_client_id TEXT,
-        related_project_id TEXT,
-        created_by_user_id TEXT,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (related_client_id) REFERENCES Clients (client_id) ON DELETE SET NULL,
-        FOREIGN KEY (related_project_id) REFERENCES Projects (project_id) ON DELETE SET NULL,
-        FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id) ON DELETE SET NULL
-    )
-    """)
-
-    # Create EmailReminders table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS EmailReminders (
-        reminder_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        scheduled_email_id INTEGER NOT NULL,
-        reminder_type TEXT NOT NULL, 
-        reminder_send_at TIMESTAMP NOT NULL, 
-        status TEXT NOT NULL DEFAULT 'pending', 
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (scheduled_email_id) REFERENCES ScheduledEmails (scheduled_email_id) ON DELETE CASCADE
-    )
-    """)
-
-    # Create ContactLists table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ContactLists (
-        list_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        list_name TEXT NOT NULL UNIQUE,
-        description TEXT,
-        created_by_user_id TEXT,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id) ON DELETE SET NULL
-    )
-    """)
-
-    # Create ContactListMembers table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ContactListMembers (
-        list_member_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        list_id INTEGER NOT NULL,
-        contact_id INTEGER NOT NULL,
-        added_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (list_id) REFERENCES ContactLists (list_id) ON DELETE CASCADE,
-        FOREIGN KEY (contact_id) REFERENCES Contacts (contact_id) ON DELETE CASCADE,
-        UNIQUE (list_id, contact_id)
-    )
-    """)
-    
-    # Create ActivityLog table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ActivityLog (
-        log_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        user_id TEXT, 
-        action_type TEXT NOT NULL, 
-        details TEXT, 
-        related_entity_type TEXT, 
-        related_entity_id TEXT, 
-        related_client_id TEXT, 
-        ip_address TEXT,
-        user_agent TEXT,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (user_id) REFERENCES Users (user_id) ON DELETE SET NULL,
-        FOREIGN KEY (related_client_id) REFERENCES Clients (client_id) ON DELETE SET NULL 
-    )
-    """)
-
-    # Create TemplateCategories table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS TemplateCategories (
-        category_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        category_name TEXT NOT NULL UNIQUE,
-        description TEXT
-    )
-    """)
-    # Pre-populate a "General" category
-    general_category_id_for_migration = None
-    try:
-        cursor.execute("INSERT OR IGNORE INTO TemplateCategories (category_name, description) VALUES (?, ?)",
-                       ('General', 'General purpose templates'))
-        # Fetch its ID for fallback during migration
-        cursor.execute("SELECT category_id FROM TemplateCategories WHERE category_name = 'General'")
-        general_row = cursor.fetchone()
-        if general_row:
-            general_category_id_for_migration = general_row[0]
-
-        # Pre-populate "Document Utilitaires" category
-        try:
-            cursor.execute("INSERT OR IGNORE INTO TemplateCategories (category_name, description) VALUES (?, ?)",
-                           ('Document Utilitaires', 'Modèles de documents utilitaires généraux (ex: catalogues, listes de prix)'))
-        except sqlite3.Error as e_cat_util:
-            print(f"Error initializing 'Document Utilitaires' category: {e_cat_util}")
-
-            # Add "Modèles Email" category
-            try:
-                cursor.execute("INSERT OR IGNORE INTO TemplateCategories (category_name, description) VALUES (?, ?)",
-                               ('Modèles Email', 'Modèles pour les corps des emails'))
-            except sqlite3.Error as e_cat_email:
-                print(f"Error initializing 'Modèles Email' category: {e_cat_email}")
-
-        conn.commit() # Commit category creation before potential DDL changes for Templates
-    except sqlite3.Error as e_cat_init: # This specifically catches errors from the 'General' category block
-        print(f"Error initializing General category or other categories in this block: {e_cat_init}")
-
-        # Decide if this is fatal or if migration can proceed without a fallback ID
-        # For now, migration will use None if this fails, which _get_or_create_category_id handles.
-
-    # Check if Templates table needs migration
-    cursor.execute("PRAGMA table_info(Templates)")
-    columns = [column[1] for column in cursor.fetchall()]
-    needs_migration = 'category' in columns and 'category_id' not in columns
-
-    if needs_migration:
-        print("Templates table needs migration. Starting migration process...")
-        try:
-            # 1. Rename Templates to Templates_old
-            cursor.execute("ALTER TABLE Templates RENAME TO Templates_old")
-            print("Renamed Templates to Templates_old.")
-
-            # 2. Create the new Templates table with category_id FOREIGN KEY
-            cursor.execute("""
-            CREATE TABLE Templates (
-                template_id INTEGER PRIMARY KEY AUTOINCREMENT,
-                template_name TEXT NOT NULL,
-                template_type TEXT NOT NULL,
-                description TEXT,
-                base_file_name TEXT,
-                language_code TEXT,
-                is_default_for_type_lang BOOLEAN DEFAULT FALSE,
-                category_id INTEGER,
-                content_definition TEXT,
-                email_subject_template TEXT,
-                email_variables_info TEXT,
-                cover_page_config_json TEXT,
-                document_mapping_config_json TEXT,
-                raw_template_file_data BLOB,
-                version TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                created_by_user_id TEXT,
-                FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id),
-                FOREIGN KEY (category_id) REFERENCES TemplateCategories(category_id) ON DELETE SET NULL,
-                UNIQUE (template_name, template_type, language_code, version)
-            )
-            """)
-            print("Created new Templates table with category_id.")
-
-            # 3. Iterate through Templates_old and insert into new Templates
-            cursor.execute("SELECT * FROM Templates_old")
-            old_templates = cursor.fetchall()
-
-            # Make sure connection's row_factory is set to sqlite3.Row for dict-like access
-            # This is usually set in get_db_connection, but ensure it's active for this part.
-            # If cursor.fetchall() returns tuples, access by index. If dicts, by key.
-            # Assuming get_db_connection sets row_factory, so fetchall() gives list of Row objects.
-
-            # Need to get the column names from Templates_old to safely map
-            cursor_old_desc = conn.execute("PRAGMA table_info(Templates_old)")
-            old_column_names = [col_info[1] for col_info in cursor_old_desc.fetchall()]
-
-            for old_template_tuple in old_templates:
-                old_template_dict = {name: val for name, val in zip(old_column_names, old_template_tuple)}
-
-                category_name_text = old_template_dict.get('category')
-                # Use the internal helper _get_or_create_category_id
-                # Pass the main cursor, not creating a new one for this helper
-                new_cat_id = _get_or_create_category_id(cursor, category_name_text, general_category_id_for_migration)
-
-                # Prepare values for insert, ensuring order and handling missing keys
-                new_template_values = (
-                    old_template_dict.get('template_id'),
-                    old_template_dict.get('template_name'),
-                    old_template_dict.get('template_type'),
-                    old_template_dict.get('description'),
-                    old_template_dict.get('base_file_name'),
-                    old_template_dict.get('language_code'),
-                    old_template_dict.get('is_default_for_type_lang', False), # Default if missing
-                    new_cat_id, # New category_id
-                    old_template_dict.get('content_definition'),
-                    old_template_dict.get('email_subject_template'),
-                    old_template_dict.get('email_variables_info'),
-                    old_template_dict.get('cover_page_config_json'),
-                    old_template_dict.get('document_mapping_config_json'),
-                    old_template_dict.get('raw_template_file_data'),
-                    old_template_dict.get('version'),
-                    old_template_dict.get('created_at'),
-                    old_template_dict.get('updated_at'),
-                    old_template_dict.get('created_by_user_id')
-                )
-
-                insert_sql = """
-                    INSERT INTO Templates (
-                        template_id, template_name, template_type, description, base_file_name,
-                        language_code, is_default_for_type_lang, category_id,
-                        content_definition, email_subject_template, email_variables_info,
-                        cover_page_config_json, document_mapping_config_json,
-                        raw_template_file_data, version, created_at, updated_at, created_by_user_id
-                    ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                """
-                cursor.execute(insert_sql, new_template_values)
-            print(f"Migrated {len(old_templates)} templates to new schema.")
-
-            # 4. Drop Templates_old
-            cursor.execute("DROP TABLE Templates_old")
-            print("Dropped Templates_old table.")
-            conn.commit()
-            print("Templates table migration completed successfully.")
-        except sqlite3.Error as e:
-            conn.rollback()
-            print(f"Error during Templates table migration: {e}. Rolled back changes.")
-            # Consider re-creating the original Templates table if migration fails critically
-            # Or, if the new Templates table was created, drop it to allow retry.
-            # For now, just rollback and log. The DB might be in an inconsistent state (e.g. Templates_old exists).
-    else:
-        # Create Templates table if it doesn't exist (fresh setup or already migrated)
-        cursor.execute("""
-        CREATE TABLE IF NOT EXISTS Templates (
-            template_id INTEGER PRIMARY KEY AUTOINCREMENT,
-            template_name TEXT NOT NULL,
-            template_type TEXT NOT NULL,
-            description TEXT,
-            base_file_name TEXT,
-            language_code TEXT,
-            is_default_for_type_lang BOOLEAN DEFAULT FALSE,
-            category_id INTEGER, -- New field
-            content_definition TEXT,
-            email_subject_template TEXT,
-            email_variables_info TEXT,
-            cover_page_config_json TEXT,
-            document_mapping_config_json TEXT,
-            raw_template_file_data BLOB,
-            version TEXT,
-            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-            updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-            created_by_user_id TEXT,
-            FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id),
-            FOREIGN KEY (category_id) REFERENCES TemplateCategories(category_id) ON DELETE SET NULL, -- Added FK
-            UNIQUE (template_name, template_type, language_code, version)
-        )
-        """)
-        # No conn.commit() here, let it be part of the larger transaction at the end of initialize_database
-
-    # Create Tasks table (New)
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS Tasks (
-        task_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        project_id TEXT NOT NULL,
-        task_name TEXT NOT NULL,
-        description TEXT,
-        status_id INTEGER,
-        assignee_team_member_id INTEGER, -- Changed to INTEGER to match TeamMembers.team_member_id
-        reporter_team_member_id INTEGER,   -- Changed to INTEGER to match TeamMembers.team_member_id
-        due_date DATE,
-        priority INTEGER DEFAULT 0,
-        estimated_hours REAL,
-        actual_hours_spent REAL,
-        parent_task_id INTEGER,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        completed_at TIMESTAMP,
-        FOREIGN KEY (project_id) REFERENCES Projects (project_id) ON DELETE CASCADE,
-        FOREIGN KEY (status_id) REFERENCES StatusSettings (status_id),
-        FOREIGN KEY (assignee_team_member_id) REFERENCES TeamMembers (team_member_id) ON DELETE SET NULL,
-        FOREIGN KEY (reporter_team_member_id) REFERENCES TeamMembers (team_member_id) ON DELETE SET NULL,
-        FOREIGN KEY (parent_task_id) REFERENCES Tasks (task_id) ON DELETE SET NULL
-    )
-    """)
-
-    # Create ClientDocuments table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ClientDocuments (
-        document_id TEXT PRIMARY KEY,
-        client_id TEXT NOT NULL,
-        project_id TEXT,
-        document_name TEXT NOT NULL,
-        file_name_on_disk TEXT NOT NULL, -- Actual name on the file system
-        file_path_relative TEXT NOT NULL, -- Relative to a base documents folder
-        document_type_generated TEXT, -- e.g., 'Proposal', 'Contract', 'Invoice'
-        source_template_id INTEGER,
-        version_tag TEXT,
-        notes TEXT,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        created_by_user_id TEXT,
-        FOREIGN KEY (client_id) REFERENCES Clients (client_id),
-        FOREIGN KEY (project_id) REFERENCES Projects (project_id),
-        FOREIGN KEY (source_template_id) REFERENCES Templates (template_id),
-        FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id)
-    )
-    """)
-
-    # Create SmtpConfigs table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS SmtpConfigs (
-        smtp_config_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        config_name TEXT NOT NULL UNIQUE,
-        smtp_server TEXT NOT NULL,
-        smtp_port INTEGER NOT NULL,
-        username TEXT,
-        password_encrypted TEXT, -- Store encrypted password
-        use_tls BOOLEAN DEFAULT TRUE,
-        is_default BOOLEAN DEFAULT FALSE,
-        sender_email_address TEXT NOT NULL,
-        sender_display_name TEXT
-    )
-    """)
-
-    # Create ApplicationSettings table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS ApplicationSettings (
-        setting_key TEXT PRIMARY KEY,
-        setting_value TEXT
-    )
-    """)
-
-    # Create KPIs table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS KPIs (
-        kpi_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        project_id TEXT NOT NULL,
-        name TEXT NOT NULL,
-        value REAL NOT NULL,
-        target REAL NOT NULL,
-        trend TEXT NOT NULL, -- 'up', 'down', 'stable'
-        unit TEXT NOT NULL,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (project_id) REFERENCES Projects (project_id) ON DELETE CASCADE
-    )
-    """)
-
-    # Create CoverPageTemplates table
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS CoverPageTemplates (
-        template_id TEXT PRIMARY KEY,
-        template_name TEXT NOT NULL UNIQUE,
-        description TEXT,
-        default_title TEXT,
-        default_subtitle TEXT,
-        default_author TEXT,
-        style_config_json TEXT, -- JSON string for detailed styling (fonts, colors, layout hints)
-        is_default_template INTEGER DEFAULT 0 NOT NULL, -- Added new flag
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        created_by_user_id TEXT,
-        FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id) ON DELETE SET NULL
-    )
-    """)
-
-    # Create CoverPages table (linking to Clients/Projects and Templates)
-    cursor.execute("""
-    CREATE TABLE IF NOT EXISTS CoverPages (
-        cover_page_id TEXT PRIMARY KEY,
-        cover_page_name TEXT, -- User-defined name for this instance
-        client_id TEXT,
-        project_id TEXT,
-        template_id TEXT,
-        title TEXT NOT NULL,
-        subtitle TEXT,
-        author_text TEXT,
-        institution_text TEXT,
-        department_text TEXT,
-        document_type_text TEXT,
-        document_version TEXT,
-        creation_date DATE, -- Renamed from document_date for consistency with table
-        logo_name TEXT,
-        logo_data BLOB,
-        custom_style_config_json TEXT,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        created_by_user_id TEXT,
-        FOREIGN KEY (client_id) REFERENCES Clients (client_id) ON DELETE SET NULL,
-        FOREIGN KEY (project_id) REFERENCES Projects (project_id) ON DELETE SET NULL,
-        FOREIGN KEY (template_id) REFERENCES CoverPageTemplates (template_id) ON DELETE SET NULL,
-        FOREIGN KEY (created_by_user_id) REFERENCES Users (user_id) ON DELETE SET NULL
-    )
-    """)
-
-    conn.commit()
-    conn.close()
-
-def get_db_connection():
-    """
-    Returns a new database connection object.
-    The connection is configured to return rows as dictionary-like objects.
-    """
-    conn = sqlite3.connect(DATABASE_NAME)
-    conn.row_factory = sqlite3.Row
-    return conn
-
-# CRUD functions for Clients
-def add_client(client_data: dict) -> str | None:
-    """
-    Adds a new client to the database.
-    Returns the new client_id if successful, otherwise None.
-    Ensures created_at and updated_at are set.
-    Expects 'category_id' instead of 'category' text.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        new_client_id = uuid.uuid4().hex
-        now = datetime.utcnow().isoformat() + "Z"
-
-        # Ensure all required fields are present, or provide defaults
-        sql = """
-            INSERT INTO Clients (
-                client_id, client_name, company_name, primary_need_description, project_identifier,
-                country_id, city_id, default_base_folder_path, status_id,
-                selected_languages, notes, category, created_at, updated_at, created_by_user_id
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            new_client_id,
-            client_data.get('client_name'),
-            client_data.get('company_name'),
-            client_data.get('primary_need_description'),
-            client_data.get('project_identifier'), # Added
-            client_data.get('country_id'),
-            client_data.get('city_id'),
-            client_data.get('default_base_folder_path'),
-            client_data.get('status_id'),
-            client_data.get('selected_languages'),
-            client_data.get('notes'),
-            client_data.get('category'),
-            now,  # created_at
-            now,  # updated_at
-            client_data.get('created_by_user_id')
-        )
-        
-        cursor.execute(sql, params)
-        conn.commit()
-        return new_client_id
-    except sqlite3.Error as e:
-        print(f"Database error in add_client: {e}")
-        # Consider raising a custom exception or logging more formally
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_client_by_id(client_id: str) -> dict | None:
-    """Retrieves a client by their ID. Returns a dict or None if not found."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Clients WHERE client_id = ?", (client_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_client_by_id: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_clients(filters: dict = None) -> list[dict]:
-    """
-    Retrieves all clients, optionally applying filters.
-    Filters can be e.g. {'status_id': 1, 'category': 'VIP'}.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        sql = "SELECT * FROM Clients"
-        params = []
-        
-        if filters:
-            where_clauses = []
-            for key, value in filters.items():
-                # Basic protection against non-column names; ideally, validate keys against known columns
-                if key in ['client_name', 'company_name', 'country_id', 'city_id', 'status_id', 'category', 'created_by_user_id']: # Add other filterable columns
-                    where_clauses.append(f"{key} = ?")
-                    params.append(value)
-            if where_clauses:
-                sql += " WHERE " + " AND ".join(where_clauses)
-                
-        cursor.execute(sql, params)
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_clients: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def update_client(client_id: str, client_data: dict) -> bool:
-    """
-    Updates an existing client's information.
-    Ensures updated_at is set to the current timestamp.
-    Returns True if update was successful, False otherwise.
-    """
-    conn = None
-    if not client_data:
-        return False # Nothing to update
-
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        now = datetime.utcnow().isoformat() + "Z"
-        client_data['updated_at'] = now
-        
-        set_clauses = []
-        params = []
-        
-        for key, value in client_data.items():
-            # Validate keys against actual column names to prevent SQL injection if keys are from unsafe source
-            # For now, assuming keys are controlled or map to valid columns
-            if key != 'client_id': # client_id should not be updated here
-                 set_clauses.append(f"{key} = ?")
-                 params.append(value)
-        
-        if not set_clauses:
-            return False # No valid fields to update
-            
-        sql = f"UPDATE Clients SET {', '.join(set_clauses)} WHERE client_id = ?"
-        params.append(client_id)
-        
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_client: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def delete_client(client_id: str) -> bool:
-    """
-    Deletes a client from the database.
-    Returns True if deletion was successful, False otherwise.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM Clients WHERE client_id = ?", (client_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_client: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def add_client_note(client_id: str, note_text: str, user_id: str = None) -> int | None:
-    """
-    Adds a new note for a client.
-    Returns the note_id if successful, otherwise None.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = """
-            INSERT INTO ClientNotes (client_id, note_text, user_id)
-            VALUES (?, ?, ?)
-        """
-        # timestamp is handled by DEFAULT CURRENT_TIMESTAMP
-        params = (client_id, note_text, user_id)
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid  # Returns the note_id of the inserted row
-    except sqlite3.Error as e:
-        print(f"Database error in add_client_note: {e}")
-        if conn:
-            conn.rollback() # Rollback any changes if an error occurred
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_client_notes(client_id: str) -> list[dict]:
-    """
-    Retrieves all notes for a given client_id, ordered by timestamp (oldest first).
-    Returns a list of dictionaries, where each dictionary represents a note.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = """
-            SELECT note_id, client_id, timestamp, note_text, user_id
-            FROM ClientNotes
-            WHERE client_id = ?
-            ORDER BY timestamp ASC
-        """
-        cursor.execute(sql, (client_id,))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_client_notes: {e}")
-        return [] # Return an empty list in case of error
-    finally:
-        if conn:
-            conn.close()
-
-# CRUD functions for Companies
-def add_company(company_data: dict) -> str | None:
-    """Adds a new company. Generates UUID for company_id. Handles created_at, updated_at."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        new_company_id = str(uuid.uuid4())
-
-        sql = """
-            INSERT INTO Companies (
-                company_id, company_name, address, payment_info, logo_path,
-                other_info, is_default, created_at, updated_at
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            new_company_id,
-            company_data.get('company_name'),
-            company_data.get('address'),
-            company_data.get('payment_info'),
-            company_data.get('logo_path'),
-            company_data.get('other_info'),
-            company_data.get('is_default', False),
-            now,  # created_at
-            now   # updated_at
-        )
-        cursor.execute(sql, params)
-        conn.commit()
-        return new_company_id
-    except sqlite3.Error as e:
-        print(f"Database error in add_company: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_company_by_id(company_id: str) -> dict | None:
-    """Fetches a company by its ID."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Companies WHERE company_id = ?", (company_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_company_by_id: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_companies() -> list[dict]:
-    """Fetches all companies."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Companies ORDER BY company_name")
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_companies: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def update_company(company_id: str, company_data: dict) -> bool:
-    """Updates company details. Manages updated_at."""
-    conn = None
-    if not company_data:
-        return False
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        company_data['updated_at'] = now
-
-        set_clauses = [f"{key} = ?" for key in company_data.keys() if key != 'company_id']
-        params = [value for key, value in company_data.items() if key != 'company_id']
-
-        if not set_clauses:
-            return False # No valid fields to update
-
-        params.append(company_id)
-        sql = f"UPDATE Companies SET {', '.join(set_clauses)} WHERE company_id = ?"
-
-        cursor.execute(sql, tuple(params))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_company: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def delete_company(company_id: str) -> bool:
-    """Deletes a company."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # ON DELETE CASCADE will handle CompanyPersonnel
-        cursor.execute("DELETE FROM Companies WHERE company_id = ?", (company_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_company: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def set_default_company(company_id: str) -> bool:
-    """Sets a company as default, ensuring only one company can be default."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        conn.isolation_level = None # Start transaction
-        cursor.execute("BEGIN")
-        # Unset other defaults
-        cursor.execute("UPDATE Companies SET is_default = FALSE WHERE is_default = TRUE AND company_id != ?", (company_id,))
-        # Set the new default
-        cursor.execute("UPDATE Companies SET is_default = TRUE WHERE company_id = ?", (company_id,))
-        conn.commit()
-        return True
-    except sqlite3.Error as e:
-        if conn:
-            conn.rollback()
-        print(f"Database error in set_default_company: {e}")
-        return False
-    finally:
-        if conn:
-            conn.isolation_level = '' # Reset isolation level
-            conn.close()
-
-def get_default_company() -> dict | None:
-    """
-    Retrieves the company marked as default.
-    Returns company data as a dictionary if found, otherwise None.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Companies WHERE is_default = TRUE")
-        row = cursor.fetchone()
-        if row:
-            return dict(row)
-        return None
-    except sqlite3.Error as e:
-        print(f"Database error in get_default_company: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-# CRUD functions for CompanyPersonnel
-def add_company_personnel(personnel_data: dict) -> int | None:
-    """Inserts new personnel linked to a company. Returns personnel_id."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        sql = """
-            INSERT INTO CompanyPersonnel (company_id, name, role, created_at)
-            VALUES (?, ?, ?, ?)
-        """
-        params = (
-            personnel_data.get('company_id'),
-            personnel_data.get('name'),
-            personnel_data.get('role'),
-            now
-        )
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.Error as e:
-        print(f"Database error in add_company_personnel: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_personnel_for_company(company_id: str, role: str = None) -> list[dict]:
-    """Fetches personnel for a company, optionally filtering by role."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = "SELECT * FROM CompanyPersonnel WHERE company_id = ?"
-        params = [company_id]
-        if role:
-            sql += " AND role = ?"
-            params.append(role)
-        sql += " ORDER BY name"
-        cursor.execute(sql, tuple(params))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_personnel_for_company: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def update_company_personnel(personnel_id: int, personnel_data: dict) -> bool:
-    """Updates personnel details."""
-    conn = None
-    if not personnel_data:
-        return False
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        # We don't update created_at, but if there was an updated_at for this table:
-        # personnel_data['updated_at'] = datetime.utcnow().isoformat() + "Z"
-
-        set_clauses = [f"{key} = ?" for key in personnel_data.keys() if key != 'personnel_id']
-        params = [value for key, value in personnel_data.items() if key != 'personnel_id']
-
-        if not set_clauses:
-            return False
-
-        params.append(personnel_id)
-        sql = f"UPDATE CompanyPersonnel SET {', '.join(set_clauses)} WHERE personnel_id = ?"
-
-        cursor.execute(sql, tuple(params))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_company_personnel: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def delete_company_personnel(personnel_id: int) -> bool:
-    """Deletes a personnel entry."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM CompanyPersonnel WHERE personnel_id = ?", (personnel_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_company_personnel: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-# CRUD functions for TemplateCategories
-def add_template_category(category_name: str, description: str = None) -> int | None:
-    """
-    Adds a new template category if it doesn't exist by name.
-    Returns the category_id of the new or existing category, or None on error.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        # Check if category already exists
-        cursor.execute("SELECT category_id FROM TemplateCategories WHERE category_name = ?", (category_name,))
-        row = cursor.fetchone()
-        if row:
-            return row['category_id']
-
-        # If not, insert new category
-        sql = "INSERT INTO TemplateCategories (category_name, description) VALUES (?, ?)"
-        cursor.execute(sql, (category_name, description))
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.Error as e:
-        print(f"Database error in add_template_category: {e}")
-        if conn:
-            conn.rollback()
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def _get_or_create_category_id(cursor: sqlite3.Cursor, category_name: str, default_category_id: int | None) -> int | None:
-    """
-    Internal helper: Gets category_id for a name, creates if not exists.
-    Uses the provided cursor and does not manage connection or transaction.
-    Returns category_id or default_category_id if name is None/empty.
-    """
-    if not category_name:
-        return default_category_id
-    try:
-        cursor.execute("SELECT category_id FROM TemplateCategories WHERE category_name = ?", (category_name,))
-        row = cursor.fetchone()
-        if row:
-            return row['category_id']
-        else:
-            # Category does not exist, create it
-            cursor.execute("INSERT INTO TemplateCategories (category_name, description) VALUES (?, ?)",
-                           (category_name, f"{category_name} (auto-created during migration)"))
-            # No conn.commit() here as it's part of a larger transaction
-            return cursor.lastrowid
-    except sqlite3.Error as e:
-        print(f"Error in _get_or_create_category_id for '{category_name}': {e}")
-        # Depending on how critical this is, you might want to raise the error
-        # or return the default_category_id as a fallback.
-        return default_category_id
-
-
-def get_template_category_by_id(category_id: int) -> dict | None:
-    """Retrieves a template category by its ID."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM TemplateCategories WHERE category_id = ?", (category_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_template_category_by_id: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_template_category_by_name(category_name: str) -> dict | None:
-    """Retrieves a template category by its name."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM TemplateCategories WHERE category_name = ?", (category_name,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_template_category_by_name: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_template_categories() -> list[dict]:
-    """Retrieves all template categories."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM TemplateCategories ORDER BY category_name")
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_template_categories: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def update_template_category(category_id: int, new_name: str = None, new_description: str = None) -> bool:
-    """
-    Updates a template category's name and/or description.
-    Returns True on success, False otherwise.
-    """
-    conn = None
-    if not new_name and not new_description:
-        return False # Nothing to update
-
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        set_clauses = []
-        params = []
-        if new_name:
-            set_clauses.append("category_name = ?")
-            params.append(new_name)
-        if new_description is not None: # Allow setting description to empty string
-            set_clauses.append("description = ?")
-            params.append(new_description)
-
-        if not set_clauses:
-            return False
-
-        sql = f"UPDATE TemplateCategories SET {', '.join(set_clauses)} WHERE category_id = ?"
-        params.append(category_id)
-
-        cursor.execute(sql, tuple(params))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_template_category: {e}")
-        if conn:
-            conn.rollback()
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def delete_template_category(category_id: int) -> bool:
-    """
-    Deletes a template category.
-    Templates using this category will have their category_id set to NULL
-    due to ON DELETE SET NULL foreign key constraint.
-    Returns True on success, False otherwise.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # Before deleting, one might want to check if it's a protected category like "General"
-        # For now, allowing deletion of any category.
-        cursor.execute("DELETE FROM TemplateCategories WHERE category_id = ?", (category_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_template_category: {e}")
-        if conn:
-            conn.rollback()
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-# CRUD functions for Templates
-def add_template(template_data: dict) -> int | None:
-    """
-    Adds a new template to the database. Returns the template_id if successful.
-    Ensures created_at and updated_at are set.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-
-        sql = """
-            INSERT INTO Templates (
-                template_name, template_type, description, base_file_name, language_code,
-                is_default_for_type_lang, category_id, content_definition, email_subject_template,
-                email_variables_info, cover_page_config_json, document_mapping_config_json,
-                raw_template_file_data, version, created_at, updated_at, created_by_user_id
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            template_data.get('template_name'),
-            template_data.get('template_type'),
-            template_data.get('description'),
-            template_data.get('base_file_name'),
-            template_data.get('language_code'),
-            template_data.get('is_default_for_type_lang', False),
-            template_data.get('category_id'), # Changed from 'category' to 'category_id'
-            template_data.get('content_definition'),
-            template_data.get('email_subject_template'),
-            template_data.get('email_variables_info'),
-            template_data.get('cover_page_config_json'),
-            template_data.get('document_mapping_config_json'),
-            template_data.get('raw_template_file_data'),
-            template_data.get('version'),
-            now,  # created_at
-            now,  # updated_at
-            template_data.get('created_by_user_id')
-        )
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid # For AUTOINCREMENT PK
-    except sqlite3.Error as e:
-        print(f"Database error in add_template: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_template_by_id(template_id: int) -> dict | None:
-    """Retrieves a template by its ID. Returns a dict or None if not found."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Templates WHERE template_id = ?", (template_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_template_by_id: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_templates_by_type(template_type: str, language_code: str = None) -> list[dict]:
-    """
-    Retrieves templates filtered by template_type.
-    If language_code is provided, also filters by language_code.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        sql = "SELECT * FROM Templates WHERE template_type = ?"
-        params = [template_type]
-        
-        if language_code:
-            sql += " AND language_code = ?"
-            params.append(language_code)
-            
-        cursor.execute(sql, tuple(params))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_templates_by_type: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def update_template(template_id: int, template_data: dict) -> bool:
-    """
-    Updates an existing template.
-    Ensures updated_at is set.
-    Returns True on success.
-    """
-    conn = None
-    if not template_data:
-        return False
-
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        template_data['updated_at'] = now
-        
-        set_clauses = []
-        params = []
-        
-        for key, value in template_data.items():
-            if key != 'template_id': # template_id should not be updated
-                set_clauses.append(f"{key} = ?")
-                params.append(value)
-        
-        if not set_clauses:
-            return False
-            
-        sql = f"UPDATE Templates SET {', '.join(set_clauses)} WHERE template_id = ?"
-        params.append(template_id)
-        
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_template: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def delete_template(template_id: int) -> bool:
-    """Deletes a template from the database. Returns True on success."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM Templates WHERE template_id = ?", (template_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_template: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def get_template_details_for_preview(template_id: int) -> dict | None:
-    """
-    Fetches base_file_name and language_code for a given template_id for preview purposes.
-    Returns a dictionary like {'base_file_name': 'name.xlsx', 'language_code': 'fr'} or None.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT base_file_name, language_code FROM Templates WHERE template_id = ?",
-            (template_id,)
-        )
-        row = cursor.fetchone()
-        if row:
-            return {'base_file_name': row['base_file_name'], 'language_code': row['language_code']}
-        return None
-    except sqlite3.Error as e:
-        print(f"Database error in get_template_details_for_preview: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_template_path_info(template_id: int) -> dict | None:
-    """
-    Fetches base_file_name (as file_name) and language_code (as language) for a given template_id.
-    Returns {'file_name': 'name.xlsx', 'language': 'fr'} or None.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute(
-            "SELECT base_file_name, language_code FROM Templates WHERE template_id = ?",
-            (template_id,)
-        )
-        row = cursor.fetchone()
-        if row:
-            return {'file_name': row['base_file_name'], 'language': row['language_code']}
-        return None
-    except sqlite3.Error as e:
-        print(f"Database error in get_template_path_info: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def delete_template_and_get_file_info(template_id: int) -> dict | None:
-    """
-    Deletes the template record by template_id after fetching its file information.
-    Returns {'file_name': 'name.xlsx', 'language': 'fr'} if successful, None otherwise.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        conn.isolation_level = None # Start transaction
-        cursor = conn.cursor()
-        cursor.execute("BEGIN")
-
-        # First, fetch the required information
-        cursor.execute(
-            "SELECT base_file_name, language_code FROM Templates WHERE template_id = ?",
-            (template_id,)
-        )
-        row = cursor.fetchone()
-
-        if not row:
-            conn.rollback()
-            print(f"Template with ID {template_id} not found for deletion.")
-            return None
-
-        file_info = {'file_name': row['base_file_name'], 'language': row['language_code']}
-
-        # Proceed with deletion
-        cursor.execute("DELETE FROM Templates WHERE template_id = ?", (template_id,))
-
-        if cursor.rowcount > 0:
-            conn.commit()
-            return file_info
-        else:
-            # This case should ideally not be reached if the select was successful
-            # but included for robustness
-            conn.rollback()
-            print(f"Failed to delete template with ID {template_id} after fetching info.")
-            return None
-
-    except sqlite3.Error as e:
-        if conn:
-            conn.rollback()
-        print(f"Database error in delete_template_and_get_file_info: {e}")
-        return None
-    finally:
-        if conn:
-            conn.isolation_level = '' # Reset isolation level
-            conn.close()
-
-def set_default_template_by_id(template_id: int) -> bool:
-    """
-    Sets a template as the default for its template_type and language_code.
-    Unsets other templates of the same type and language.
-    Returns True on success, False on error.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        conn.isolation_level = None # Start transaction
-        cursor = conn.cursor()
-        cursor.execute("BEGIN")
-
-        # Get template_type and language_code for the given template_id
-        cursor.execute(
-            "SELECT template_type, language_code FROM Templates WHERE template_id = ?",
-            (template_id,)
-        )
-        template_info = cursor.fetchone()
-
-        if not template_info:
-            print(f"Template with ID {template_id} not found.")
-            conn.rollback()
-            return False
-
-        current_template_type = template_info['template_type']
-        current_language_code = template_info['language_code']
-
-        # Set is_default_for_type_lang = 0 for all templates of the same type and language
-        cursor.execute(
-            """
-            UPDATE Templates
-            SET is_default_for_type_lang = 0
-            WHERE template_type = ? AND language_code = ?
-            """,
-            (current_template_type, current_language_code)
-        )
-
-        # Set is_default_for_type_lang = 1 for the specified template_id
-        cursor.execute(
-            "UPDATE Templates SET is_default_for_type_lang = 1 WHERE template_id = ?",
-            (template_id,)
-        )
-
-        conn.commit()
-        return True
-
-    except sqlite3.Error as e:
-        if conn:
-            conn.rollback()
-        print(f"Database error in set_default_template_by_id: {e}")
-        return False
-    finally:
-        if conn:
-            conn.isolation_level = '' # Reset isolation level
-            conn.close()
-
-def add_default_template_if_not_exists(template_data: dict) -> int | None:
-    """
-    Adds a template to the Templates table if it doesn't already exist
-    based on template_name, template_type, and language_code.
-    Returns the template_id of the new or existing template, or None on error.
-    Expects template_data to include:
-        'template_name' (e.g., "Proforma"),
-        'template_type' (e.g., "document_excel", "document_word"),
-        'language_code' (e.g., "fr", "en"),
-        'base_file_name' (e.g., "proforma_template.xlsx"),
-        'description' (optional),
-        'category' (optional, e.g., "Finance", "Technical"),
-        'is_default_for_type_lang' (optional, boolean, defaults to False),
-        'category_name' (optional, string, defaults to "General")
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor() # Get a cursor from the connection
-
-        name = template_data.get('template_name')
-        ttype = template_data.get('template_type')
-        lang = template_data.get('language_code')
-        filename = template_data.get('base_file_name')
-        category_name_text = template_data.get('category_name', "General")
-
-        if not all([name, ttype, lang, filename]):
-            print(f"Error: Missing required fields for default template: {template_data}")
-            return None
-
-        # Get or create category_id (using a separate connection for this, or pass cursor)
-        # For simplicity here, calling the public function.
-        # In a high-performance scenario, might pass the cursor.
-        category_id = add_template_category(category_name_text, f"{category_name_text} (auto-created)")
-        if category_id is None:
-            print(f"Error: Could not get or create category_id for '{category_name_text}'.")
-            return None # Cannot proceed without a category_id
-
-        # Check if this specific template (name, type, lang) already exists
-        cursor.execute("""
-            SELECT template_id FROM Templates
-            WHERE template_name = ? AND template_type = ? AND language_code = ?
-        """, (name, ttype, lang))
-        existing_template = cursor.fetchone()
-
-        if existing_template:
-            print(f"Default template '{name}' ({ttype}, {lang}) already exists with ID: {existing_template['template_id']}.")
-            return existing_template['template_id']
-        else:
-            now = datetime.utcnow().isoformat() + "Z"
-            sql = """
-                INSERT INTO Templates (
-                    template_name, template_type, language_code, base_file_name,
-                    description, category_id, is_default_for_type_lang,
-                    email_subject_template, -- Added email_subject_template
-                    created_at, updated_at
-                    -- created_by_user_id could be NULL or a system user ID
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?) -- Added placeholder for email_subject_template
-            """
-            params = (
-                name,
-                ttype,
-                lang,
-                filename,
-                template_data.get('description', f"Default {name} template"),
-                category_id, # Use the fetched/created category_id
-                template_data.get('is_default_for_type_lang', True),
-                template_data.get('email_subject_template'), # Get email_subject_template
-                now,
-                now
-            )
-            cursor.execute(sql, params)
-            conn.commit()
-            new_id = cursor.lastrowid
-            print(f"Added default template '{name}' ({ttype}, {lang}) with Category ID: {category_id}, new Template ID: {new_id}.")
-            return new_id
-
-    except sqlite3.Error as e:
-        print(f"Database error in add_default_template_if_not_exists for '{template_data.get('template_name')}': {e}")
-        if conn:
-            conn.rollback() # Rollback on error
-        return None
-    finally:
-        if conn:
-            conn.close()
+from db.db_config import DATABASE_NAME, APP_ROOT_DIR_CONTEXT, LOGO_SUBDIR_CONTEXT, get_db_connection
+from db.db_schema import initialize_database
+from db.db_crud_users import (
+    add_user, get_user_by_id, get_user_by_username, update_user, verify_user_password, delete_user,
+    add_team_member, get_team_member_by_id, get_all_team_members, update_team_member, delete_team_member,
+    add_company, get_company_by_id, get_all_companies, update_company, delete_company, set_default_company, get_default_company,
+    add_company_personnel, get_personnel_for_company, update_company_personnel, delete_company_personnel,
+    add_smtp_config, get_smtp_config_by_id, get_default_smtp_config, get_all_smtp_configs,
+    update_smtp_config, delete_smtp_config, set_default_smtp_config,
+    get_setting, set_setting
+)
+from db.db_crud_clients import (
+    add_client, get_client_by_id, get_all_clients, update_client, delete_client,
+    add_client_note, get_client_notes,
+    add_contact, get_contact_by_id, get_contact_by_email, get_all_contacts, update_contact, delete_contact,
+    link_contact_to_client, unlink_contact_from_client, get_contacts_for_client, get_clients_for_contact,
+    get_specific_client_contact_link_details, update_client_contact_link,
+    get_all_countries, get_country_by_id, get_country_by_name, add_country,
+    get_all_cities, add_city, get_city_by_name_and_country_id, get_city_by_id
+)
+from db.db_crud_projects import (
+    add_project, get_project_by_id, get_projects_by_client_id, get_all_projects, update_project, delete_project,
+    add_task, get_task_by_id, get_tasks_by_project_id, update_task, delete_task,
+    add_kpi, get_kpi_by_id, get_kpis_for_project, update_kpi, delete_kpi,
+    get_all_status_settings, get_status_setting_by_id, get_status_setting_by_name,
+    get_all_tasks, get_tasks_by_assignee_id
+)
+from db.db_crud_products import (
+    add_product, get_product_by_id, get_product_by_name, get_all_products,
+    update_product, delete_product, add_product_equivalence,
+    get_equivalent_products, get_all_product_equivalencies, remove_product_equivalence,
+    add_product_to_client_or_project, get_products_for_client_or_project,
+    update_client_project_product, remove_product_from_client_or_project,
+    get_products_by_name_pattern, get_all_products_for_selection,
+    get_all_products_for_selection_filtered
+)
+from db.db_crud_templates import (
+    add_template_category, get_template_category_by_id, get_template_category_by_name,
+    get_all_template_categories, update_template_category, delete_template_category,
+    add_template, get_template_by_id, get_templates_by_type, update_template,
+    delete_template, add_client_document, get_document_by_id,
+    get_documents_for_client, get_documents_for_project, update_client_document,
+    delete_client_document, get_template_details_for_preview,
+    get_template_path_info, delete_template_and_get_file_info,
+    set_default_template_by_id, add_default_template_if_not_exists,
+    get_distinct_languages_for_template_type, get_all_file_based_templates,
+    get_templates_by_category_id
+)
 
 # CRUD functions for Projects
 def add_project(project_data: dict) -> str | None:
@@ -2131,1287 +385,6 @@ def delete_task(task_id: int) -> bool:
         if conn:
             conn.close()
 
-# CRUD functions for Users
-def add_user(user_data: dict) -> str | None:
-    """
-    Adds a new user to the database.
-    Generates user_id (UUID), hashes password.
-    Returns the new user_id if successful, otherwise None.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        new_user_id = uuid.uuid4().hex
-        now = datetime.utcnow().isoformat() + "Z"
-        
-        if 'password' not in user_data or not user_data['password']:
-            print("Password is required to create a user.")
-            return None
-        if 'username' not in user_data or not user_data['username']:
-            print("Username is required to create a user.")
-            return None
-        if 'email' not in user_data or not user_data['email']:
-            print("Email is required to create a user.")
-            return None
-        if 'role' not in user_data or not user_data['role']:
-            print("Role is required to create a user.")
-            return None
-
-        password_hash = hashlib.sha256(user_data['password'].encode('utf-8')).hexdigest()
-
-        sql = """
-            INSERT INTO Users (
-                user_id, username, password_hash, full_name, email, role, 
-                is_active, created_at, updated_at, last_login_at
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            new_user_id,
-            user_data.get('username'),
-            password_hash,
-            user_data.get('full_name'),
-            user_data.get('email'),
-            user_data.get('role'),
-            user_data.get('is_active', True),
-            now,  # created_at
-            now,  # updated_at
-            user_data.get('last_login_at') 
-        )
-        
-        cursor.execute(sql, params)
-        conn.commit()
-        return new_user_id
-    except sqlite3.Error as e:
-        print(f"Database error in add_user: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_user_by_id(user_id: str) -> dict | None:
-    """Retrieves a user by their ID."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Users WHERE user_id = ?", (user_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_user_by_id: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_user_by_username(username: str) -> dict | None:
-    """Retrieves a user by their username."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Users WHERE username = ?", (username,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_user_by_username: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def update_user(user_id: str, user_data: dict) -> bool:
-    """
-    Updates an existing user's information.
-    If 'password' is in user_data, it will be hashed and updated.
-    Sets updated_at. Returns True on success.
-    """
-    conn = None
-    if not user_data:
-        return False
-
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        now = datetime.utcnow().isoformat() + "Z"
-        user_data['updated_at'] = now
-        
-        if 'password' in user_data:
-            if user_data['password']: # Ensure password is not empty
-                user_data['password_hash'] = hashlib.sha256(user_data.pop('password').encode('utf-8')).hexdigest()
-            else:
-                user_data.pop('password') # Remove empty password from update data
-        
-        set_clauses = []
-        params = []
-        
-        valid_columns = ['username', 'password_hash', 'full_name', 'email', 'role', 'is_active', 'updated_at', 'last_login_at']
-        for key, value in user_data.items():
-            if key in valid_columns:
-                 set_clauses.append(f"{key} = ?")
-                 params.append(value)
-        
-        if not set_clauses: # No valid fields to update (e.g. only empty password was provided)
-            return False 
-            
-        sql = f"UPDATE Users SET {', '.join(set_clauses)} WHERE user_id = ?"
-        params.append(user_id)
-        
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_user: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def verify_user_password(username: str, password: str) -> dict | None:
-    """
-    Verifies a user's password.
-    Returns user data (dict) if verification is successful, otherwise None.
-    """
-    user = get_user_by_username(username)
-    if user and user['is_active']: # Check if user exists and is active
-        password_hash = hashlib.sha256(password.encode('utf-8')).hexdigest()
-        if password_hash == user['password_hash']:
-            # Optionally update last_login_at here if desired
-            # update_user(user['user_id'], {'last_login_at': datetime.utcnow().isoformat() + "Z"})
-            return user
-    return None
-
-def delete_user(user_id: str) -> bool:
-    """
-    Deletes a user (hard delete).
-    Returns True if deletion was successful.
-    Note: TeamMembers.user_id will be set to NULL due to ON DELETE SET NULL.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM Users WHERE user_id = ?", (user_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_user: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-# CRUD functions for TeamMembers
-def add_team_member(member_data: dict) -> int | None:
-    """
-    Adds a new team member. Returns team_member_id (AUTOINCREMENT) or None.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-
-        sql = """
-            INSERT INTO TeamMembers (
-                user_id, full_name, email, role_or_title, department, 
-                phone_number, profile_picture_url, is_active, notes, 
-                hire_date, performance, skills,
-                created_at, updated_at
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            member_data.get('user_id'), # Can be None
-            member_data.get('full_name'),
-            member_data.get('email'),
-            member_data.get('role_or_title'),
-            member_data.get('department'),
-            member_data.get('phone_number'),
-            member_data.get('profile_picture_url'),
-            member_data.get('is_active', True),
-            member_data.get('notes'),
-            member_data.get('hire_date'),
-            member_data.get('performance', 0),
-            member_data.get('skills'),
-            now, # created_at
-            now  # updated_at
-        )
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.Error as e:
-        print(f"Database error in add_team_member: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_team_member_by_id(team_member_id: int) -> dict | None:
-    """Retrieves a team member by their ID."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM TeamMembers WHERE team_member_id = ?", (team_member_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_team_member_by_id: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_team_members(filters: dict = None) -> list[dict]:
-    """
-    Retrieves all team members, optionally applying filters.
-    Allowed filters: is_active (boolean), department (string).
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        sql = "SELECT * FROM TeamMembers"
-        params = []
-        
-        if filters:
-            where_clauses = []
-            allowed_filters = ['is_active', 'department', 'user_id'] 
-            for key, value in filters.items():
-                if key in allowed_filters:
-                    if key == 'is_active' and isinstance(value, bool):
-                         where_clauses.append(f"{key} = ?")
-                         params.append(1 if value else 0)
-                    else:
-                        where_clauses.append(f"{key} = ?")
-                        params.append(value)
-            if where_clauses:
-                sql += " WHERE " + " AND ".join(where_clauses)
-                
-        cursor.execute(sql, params)
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_team_members: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def update_team_member(team_member_id: int, member_data: dict) -> bool:
-    """
-    Updates an existing team member. Sets updated_at.
-    Returns True on success.
-    """
-    conn = None
-    if not member_data:
-        return False
-
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        member_data['updated_at'] = now
-        
-        set_clauses = []
-        params = []
-        
-        valid_columns = [
-            'user_id', 'full_name', 'email', 'role_or_title', 'department', 
-            'phone_number', 'profile_picture_url', 'is_active', 'notes',
-            'hire_date', 'performance', 'skills', 'updated_at'
-        ]
-        for key, value in member_data.items():
-            if key in valid_columns:
-                set_clauses.append(f"{key} = ?")
-                params.append(value)
-        
-        if not set_clauses:
-            return False
-            
-        sql = f"UPDATE TeamMembers SET {', '.join(set_clauses)} WHERE team_member_id = ?"
-        params.append(team_member_id)
-        
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_team_member: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-def delete_team_member(team_member_id: int) -> bool:
-    """Deletes a team member (hard delete). Returns True on success."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM TeamMembers WHERE team_member_id = ?", (team_member_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_team_member: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-# CRUD functions for Contacts
-def add_contact(contact_data: dict) -> int | None:
-    """Adds a new contact. Returns contact_id (AUTOINCREMENT) or None."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-
-        # Fallback for 'name' if 'displayName' is not provided
-        name_to_insert = contact_data.get('displayName', contact_data.get('name'))
-
-        sql = """
-            INSERT INTO Contacts (
-                name, email, phone, position, company_name, notes,
-                givenName, familyName, displayName, phone_type, email_type,
-                address_formattedValue, address_streetAddress, address_city,
-                address_region, address_postalCode, address_country,
-                organization_name, organization_title, birthday_date,
-                created_at, updated_at
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            name_to_insert,
-            contact_data.get('email'),
-            contact_data.get('phone'),
-            contact_data.get('position'),
-            contact_data.get('company_name'),
-            contact_data.get('notes'), # notes_text
-            contact_data.get('givenName'),
-            contact_data.get('familyName'),
-            contact_data.get('displayName'),
-            contact_data.get('phone_type'),
-            contact_data.get('email_type'),
-            contact_data.get('address_formattedValue'),
-            contact_data.get('address_streetAddress'),
-            contact_data.get('address_city'),
-            contact_data.get('address_region'),
-            contact_data.get('address_postalCode'),
-            contact_data.get('address_country'),
-            contact_data.get('organization_name'),
-            contact_data.get('organization_title'),
-            contact_data.get('birthday_date'),
-            now, now
-        )
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.Error as e:
-        print(f"Database error in add_contact: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_contact_by_id(contact_id: int) -> dict | None:
-    """Retrieves a contact by their ID."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Contacts WHERE contact_id = ?", (contact_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_contact_by_id: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_contact_by_email(email: str) -> dict | None:
-    """Retrieves a contact by their email."""
-    conn = None
-    if not email: return None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Contacts WHERE email = ?", (email,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_contact_by_email: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_all_contacts(filters: dict = None) -> list[dict]:
-    """
-    Retrieves all contacts. Filters by 'company_name' (exact) or 'name' (partial LIKE).
-    Now selects all new columns.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = "SELECT * FROM Contacts" # Selects all columns including new ones
-        params = []
-        where_clauses = []
-        if filters:
-            if 'company_name' in filters:
-                where_clauses.append("company_name = ?")
-                params.append(filters['company_name'])
-            if 'name' in filters: # This will search in 'name' or 'displayName'
-                where_clauses.append("(name LIKE ? OR displayName LIKE ?)")
-                params.append(f"%{filters['name']}%")
-                params.append(f"%{filters['name']}%")
-            # TODO: Add filters for new fields if needed
-        
-        if where_clauses:
-            sql += " WHERE " + " AND ".join(where_clauses)
-            
-        cursor.execute(sql, params)
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_contacts: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def update_contact(contact_id: int, contact_data: dict) -> bool:
-    """Updates an existing contact. Sets updated_at."""
-    conn = None
-    if not contact_data: return False
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        contact_data['updated_at'] = now
-
-        # Fallback for 'name' if 'displayName' is provided and 'name' is not
-        if 'displayName' in contact_data and 'name' not in contact_data:
-            contact_data['name'] = contact_data['displayName']
-        elif 'name' in contact_data and 'displayName' not in contact_data:
-             # If only 'name' is provided, ensure 'displayName' is also updated if it's meant to be the primary display
-             # This depends on application logic, for now, we'll update 'name' if 'displayName' isn't explicitly set to something else.
-             # A more robust approach might be to always set displayName = name if displayName is not in contact_data.
-             # For now, let's assume if 'displayName' is not in contact_data, it's not being changed.
-             pass
-
-
-        # Ensure only valid columns are updated
-        valid_columns = [
-            'name', 'email', 'phone', 'position', 'company_name', 'notes',
-            'givenName', 'familyName', 'displayName', 'phone_type', 'email_type',
-            'address_formattedValue', 'address_streetAddress', 'address_city',
-            'address_region', 'address_postalCode', 'address_country',
-            'organization_name', 'organization_title', 'birthday_date', 'updated_at'
-        ]
-
-        set_clauses = []
-        params = []
-
-        for key, value in contact_data.items():
-            if key in valid_columns: # Check if the key is a valid column to update
-                set_clauses.append(f"{key} = ?")
-                params.append(value)
-            elif key == 'contact_id': # Skip primary key
-                continue
-        
-        if not set_clauses:
-            print("Warning: No valid fields to update in update_contact.")
-            return False
-
-        params.append(contact_id)
-        
-        sql = f"UPDATE Contacts SET {', '.join(set_clauses)} WHERE contact_id = ?"
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_contact: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-def delete_contact(contact_id: int) -> bool:
-    """Deletes a contact. Associated ClientContacts are handled by ON DELETE CASCADE."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM Contacts WHERE contact_id = ?", (contact_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_contact: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-# Functions for ClientContacts association
-def link_contact_to_client(client_id: str, contact_id: int, is_primary: bool = False, can_receive_documents: bool = True) -> int | None:
-    """Links a contact to a client."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = """
-            INSERT INTO ClientContacts (client_id, contact_id, is_primary_for_client, can_receive_documents)
-            VALUES (?, ?, ?, ?)
-        """
-        params = (client_id, contact_id, is_primary, can_receive_documents)
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.Error as e: # Handles UNIQUE constraint violation if link already exists
-        print(f"Database error in link_contact_to_client: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def unlink_contact_from_client(client_id: str, contact_id: int) -> bool:
-    """Unlinks a contact from a client."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = "DELETE FROM ClientContacts WHERE client_id = ? AND contact_id = ?"
-        cursor.execute(sql, (client_id, contact_id))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in unlink_contact_from_client: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-def get_contacts_for_client(client_id: str) -> list[dict]:
-    """Retrieves all contacts for a given client, including link details."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = """
-            SELECT c.*, cc.is_primary_for_client, cc.can_receive_documents, cc.client_contact_id
-            FROM Contacts c
-            JOIN ClientContacts cc ON c.contact_id = cc.contact_id
-            WHERE cc.client_id = ?
-        """
-        cursor.execute(sql, (client_id,))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_contacts_for_client: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def get_clients_for_contact(contact_id: int) -> list[dict]:
-    """Retrieves all clients associated with a contact."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = """
-            SELECT cl.*, cc.is_primary_for_client, cc.can_receive_documents, cc.client_contact_id
-            FROM Clients cl
-            JOIN ClientContacts cc ON cl.client_id = cc.client_id
-            WHERE cc.contact_id = ?
-        """
-        cursor.execute(sql, (contact_id,))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_clients_for_contact: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def get_specific_client_contact_link_details(client_id: str, contact_id: int) -> dict | None:
-    """
-    Retrieves specific details (client_contact_id, is_primary_for_client, can_receive_documents)
-    for a single link between a client and a contact.
-    Returns a dict if the link is found, otherwise None.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = """
-            SELECT client_contact_id, is_primary_for_client, can_receive_documents
-            FROM ClientContacts
-            WHERE client_id = ? AND contact_id = ?
-        """
-        cursor.execute(sql, (client_id, contact_id))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_specific_client_contact_link_details: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def update_client_contact_link(client_contact_id: int, details: dict) -> bool:
-    """Updates details of a client-contact link (is_primary, can_receive_documents)."""
-    conn = None
-    if not details or not any(key in details for key in ['is_primary_for_client', 'can_receive_documents']):
-        return False
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        set_clauses = []
-        params = []
-        if 'is_primary_for_client' in details:
-            set_clauses.append("is_primary_for_client = ?")
-            params.append(details['is_primary_for_client'])
-        if 'can_receive_documents' in details:
-            set_clauses.append("can_receive_documents = ?")
-            params.append(details['can_receive_documents'])
-        
-        if not set_clauses: return False # Should not happen due to check above
-
-        params.append(client_contact_id)
-        sql = f"UPDATE ClientContacts SET {', '.join(set_clauses)} WHERE client_contact_id = ?"
-        
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_client_contact_link: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-def get_all_product_equivalencies() -> list[dict]:
-    """
-    Retrieves all product equivalencies with product details for both products in the pair.
-    Returns a list of dictionaries.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = """
-            SELECT
-                pe.equivalence_id,
-                pe.product_id_a,
-                pA.product_name AS product_name_a,
-                pA.language_code AS language_code_a,
-                pA.weight AS weight_a,
-                pA.dimensions AS dimensions_a,
-                pe.product_id_b,
-                pB.product_name AS product_name_b,
-                pB.language_code AS language_code_b,
-                pB.weight AS weight_b,
-                pB.dimensions AS dimensions_b
-            FROM ProductEquivalencies pe
-            JOIN Products pA ON pe.product_id_a = pA.product_id
-            JOIN Products pB ON pe.product_id_b = pB.product_id
-            ORDER BY pA.product_name, pB.product_name;
-        """
-        cursor.execute(sql)
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_product_equivalencies: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def remove_product_equivalence(equivalence_id: int) -> bool:
-    """
-    Deletes a product equivalence by its equivalence_id.
-    Returns True if deletion was successful, False otherwise.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM ProductEquivalencies WHERE equivalence_id = ?", (equivalence_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in remove_product_equivalence: {e}")
-        return False
-    finally:
-        if conn:
-            conn.close()
-
-
-# CRUD functions for ProductEquivalencies
-def add_product_equivalence(product_id_a: int, product_id_b: int) -> int | None:
-    """
-    Adds a product equivalence pair.
-    Ensures product_id_a < product_id_b to maintain uniqueness.
-    Returns equivalence_id of the new or existing record.
-    """
-    if product_id_a == product_id_b:
-        print("Error: Cannot create equivalence for a product with itself.")
-        return None
-
-    p_a = min(product_id_a, product_id_b)
-    p_b = max(product_id_a, product_id_b)
-
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = "INSERT INTO ProductEquivalencies (product_id_a, product_id_b) VALUES (?, ?)"
-        cursor.execute(sql, (p_a, p_b))
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.IntegrityError: # Pair already exists
-        print(f"IntegrityError: Product equivalence pair ({p_a}, {p_b}) likely already exists.")
-        cursor.execute("SELECT equivalence_id FROM ProductEquivalencies WHERE product_id_a = ? AND product_id_b = ?", (p_a, p_b))
-        row = cursor.fetchone()
-        return row['equivalence_id'] if row else None # Should find it if IntegrityError was due to this UNIQUE constraint
-    except sqlite3.Error as e:
-        print(f"Database error in add_product_equivalence: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_equivalent_products(product_id: int) -> list[dict]:
-    """
-    Retrieves all products equivalent to the given product_id.
-    Returns a list of product dictionaries.
-    """
-    conn = None
-    equivalent_product_ids = set()
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        # Find pairs where product_id is product_id_a
-        cursor.execute("SELECT product_id_b FROM ProductEquivalencies WHERE product_id_a = ?", (product_id,))
-        rows_a = cursor.fetchall()
-        for row in rows_a:
-            equivalent_product_ids.add(row['product_id_b'])
-
-        # Find pairs where product_id is product_id_b
-        cursor.execute("SELECT product_id_a FROM ProductEquivalencies WHERE product_id_b = ?", (product_id,))
-        rows_b = cursor.fetchall()
-        for row in rows_b:
-            equivalent_product_ids.add(row['product_id_a'])
-
-        equivalent_products = []
-        if equivalent_product_ids:
-            # Fetch full details for each equivalent product ID
-            # Using a placeholder for multiple IDs to avoid complex IN clause for now, iterate instead.
-            # For larger sets, an IN clause would be more efficient.
-            for eq_id in equivalent_product_ids:
-                prod_details = get_product_by_id(eq_id) # This function now includes weight and dimensions
-                if prod_details:
-                    equivalent_products.append(prod_details)
-        return equivalent_products
-
-    except sqlite3.Error as e:
-        print(f"Database error in get_equivalent_products: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-# CRUD functions for ProductEquivalencies
-def add_product_equivalence(product_id_a: int, product_id_b: int) -> int | None:
-    """
-    Adds a product equivalence pair.
-    Ensures product_id_a < product_id_b to maintain uniqueness.
-    Returns equivalence_id of the new or existing record.
-    """
-    if product_id_a == product_id_b:
-        print("Error: Cannot create equivalence for a product with itself.")
-        return None
-
-    # Ensure p_a is always the smaller ID
-    p_a = min(product_id_a, product_id_b)
-    p_b = max(product_id_a, product_id_b)
-
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = "INSERT INTO ProductEquivalencies (product_id_a, product_id_b) VALUES (?, ?)"
-        cursor.execute(sql, (p_a, p_b))
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.IntegrityError: # Pair already exists
-        print(f"IntegrityError: Product equivalence pair ({p_a}, {p_b}) likely already exists.")
-        # Fetch the ID of the existing pair
-        try:
-            cursor.execute("SELECT equivalence_id FROM ProductEquivalencies WHERE product_id_a = ? AND product_id_b = ?", (p_a, p_b))
-            row = cursor.fetchone()
-            if row:
-                return row['equivalence_id']
-            else:
-                # This case should be rare if IntegrityError was due to the unique constraint
-                print(f"Warning: IntegrityError for pair ({p_a}, {p_b}) but could not retrieve existing ID.")
-                return None
-        except sqlite3.Error as e_select:
-            print(f"Database error while trying to retrieve existing equivalence_id for ({p_a}, {p_b}): {e_select}")
-            return None
-    except sqlite3.Error as e:
-        print(f"Database error in add_product_equivalence: {e}")
-        if conn:
-            conn.rollback() # Rollback if not an integrity error on insert
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_equivalent_products(product_id: int) -> list[dict]:
-    """
-    Retrieves all products equivalent to the given product_id.
-    Returns a list of product dictionaries (including weight and dimensions).
-    """
-    conn = None
-    equivalent_product_ids = set()
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        # Find pairs where product_id is product_id_a
-        cursor.execute("SELECT product_id_b FROM ProductEquivalencies WHERE product_id_a = ?", (product_id,))
-        rows_a = cursor.fetchall()
-        for row in rows_a:
-            equivalent_product_ids.add(row['product_id_b'])
-
-        # Find pairs where product_id is product_id_b
-        cursor.execute("SELECT product_id_a FROM ProductEquivalencies WHERE product_id_b = ?", (product_id,))
-        rows_b = cursor.fetchall()
-        for row in rows_b:
-            equivalent_product_ids.add(row['product_id_a'])
-
-        equivalent_products_details = []
-        if equivalent_product_ids:
-            for eq_id in equivalent_product_ids:
-                prod_details = get_product_by_id(eq_id) # get_product_by_id will be updated to include new fields
-                if prod_details:
-                    equivalent_products_details.append(prod_details)
-        return equivalent_products_details
-
-    except sqlite3.Error as e:
-        print(f"Database error in get_equivalent_products: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-# CRUD functions for Products
-def add_product(product_data: dict) -> int | None:
-    """Adds a new product, including weight and dimensions. Returns product_id or None."""
-    conn = None
-    try:
-        product_name = product_data.get('product_name')
-        base_unit_price = product_data.get('base_unit_price')
-        language_code = product_data.get('language_code', 'fr') # Default to 'fr'
-
-        if not product_name:
-            print("Error in add_product: 'product_name' is required.")
-            # Consider raising an error or returning a more specific indicator of failure
-            return None
-        if base_unit_price is None: # Price can be 0.0, so check for None explicitly
-            print("Error in add_product: 'base_unit_price' is required.")
-            # Consider raising an error or returning a more specific indicator of failure
-            return None
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        sql = """
-            INSERT INTO Products (
-                product_name, description, category, language_code, base_unit_price,
-                unit_of_measure, weight, dimensions, is_active, created_at, updated_at
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            product_name, product_data.get('description'),
-            product_data.get('category'),
-            language_code,
-            base_unit_price,
-            product_data.get('unit_of_measure'),
-            product_data.get('weight'),
-            product_data.get('dimensions'),
-            product_data.get('is_active', True),
-            now, now
-        )
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid
-    except sqlite3.IntegrityError as e:
-        # This will catch the UNIQUE constraint violation for (product_name, language_code)
-        print(f"IntegrityError in add_product for name '{product_name}' and lang '{language_code}': {e}")
-        if conn:
-            conn.rollback()
-        # Optionally, here you could query for the existing product_id if needed:
-        # cursor.execute("SELECT product_id FROM Products WHERE product_name = ? AND language_code = ?", (product_name, language_code))
-        # existing_product = cursor.fetchone()
-        # if existing_product: return existing_product['product_id']
-        return None # Indicates "add failed" or "already exists"
-    except sqlite3.Error as e:
-        print(f"Database error in add_product: {e}")
-        if conn:
-            conn.rollback()
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_product_by_id(product_id: int) -> dict | None:
-    """Retrieves a product by ID, including weight and dimensions."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # Ensure all desired columns including new ones are selected
-        cursor.execute("SELECT product_id, product_name, description, category, language_code, base_unit_price, unit_of_measure, weight, dimensions, is_active, created_at, updated_at FROM Products WHERE product_id = ?", (product_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_product_by_id: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_product_by_name(product_name: str) -> dict | None:
-    """Retrieves a product by its exact name, including weight and dimensions. Returns a dict or None if not found."""
-    conn = None
-    if not product_name:
-        return None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # Ensure all desired columns including new ones are selected
-        cursor.execute("SELECT product_id, product_name, description, category, language_code, base_unit_price, unit_of_measure, weight, dimensions, is_active, created_at, updated_at FROM Products WHERE product_name = ?", (product_name,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_product_by_name: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_products(filters: dict = None) -> list[dict]:
-    """Retrieves all products, including weight and dimensions. Filters by category (exact) or product_name (partial LIKE)."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # Ensure all desired columns including new ones are selected
-        sql = "SELECT product_id, product_name, description, category, language_code, base_unit_price, unit_of_measure, weight, dimensions, is_active, created_at, updated_at FROM Products"
-        params = []
-        where_clauses = []
-        if filters:
-            if 'category' in filters:
-                where_clauses.append("category = ?")
-                params.append(filters['category'])
-            if 'product_name' in filters:
-                where_clauses.append("product_name LIKE ?")
-                params.append(f"%{filters['product_name']}%")
-        
-        if where_clauses:
-            sql += " WHERE " + " AND ".join(where_clauses)
-            
-        cursor.execute(sql, params)
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_products: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def update_product(product_id: int, product_data: dict) -> bool:
-    """Updates an existing product, including weight and dimensions. Sets updated_at."""
-    conn = None
-    if not product_data:
-        return False
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        now = datetime.utcnow().isoformat() + "Z"
-        product_data['updated_at'] = now
-        
-        # Ensure only valid columns are updated
-        valid_columns = [
-            'product_name', 'description', 'category', 'language_code',
-            'base_unit_price', 'unit_of_measure', 'weight', 'dimensions',
-            'is_active', 'updated_at'
-        ]
-        
-        set_clauses = []
-        params = []
-        for key, value in product_data.items():
-            if key in valid_columns: # Check if the key is a valid column to update
-                set_clauses.append(f"{key} = ?")
-                params.append(value)
-
-        if not set_clauses:
-            print("Warning: No valid fields to update in update_product.")
-            return False
-
-        params.append(product_id)
-        sql = f"UPDATE Products SET {', '.join(set_clauses)} WHERE product_id = ?"
-
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_product: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-def delete_product(product_id: int) -> bool:
-    """Deletes a product. Associated ClientProjectProducts are handled by ON DELETE CASCADE."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("DELETE FROM Products WHERE product_id = ?", (product_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_product: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-def get_products_by_name_pattern(pattern: str) -> list[dict] | None:
-    """
-    Retrieves products where the product_name matches the given pattern (LIKE %pattern%).
-    Returns a list of up to 10 matching products as dictionaries, or None on error.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        search_pattern = f"%{pattern}%"
-        # Ensure all desired columns including new ones are selected
-        sql = """
-            SELECT product_id, product_name, description, category, language_code, base_unit_price, unit_of_measure, weight, dimensions, is_active, created_at, updated_at
-            FROM Products
-            WHERE product_name LIKE ?
-            ORDER BY product_name
-            LIMIT 10
-        """
-        cursor.execute(sql, (search_pattern,))
-        rows = cursor.fetchall()
-
-        products = [dict(row) for row in rows]
-        return products
-
-    except sqlite3.Error as e:
-        print(f"Database error in get_products_by_name_pattern: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_products_for_selection(language_code: str = None, name_pattern: str = None) -> list[dict]:
-    """
-    Retrieves active products for selection, optionally filtered by language_code
-    and/or name_pattern (searches product_name and description).
-    Orders by product_name.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        sql_params = []
-        # Start with base selection of active products
-        sql_where_clauses = ["is_active = TRUE"]
-
-        if language_code:
-            sql_where_clauses.append("language_code = ?")
-            sql_params.append(language_code)
-
-        if name_pattern:
-            # The name_pattern should be passed with wildcards already, e.g., "%search_term%"
-            sql_where_clauses.append("(product_name LIKE ? OR description LIKE ?)")
-            sql_params.append(name_pattern)
-            sql_params.append(name_pattern)
-
-        sql = f"""
-            SELECT product_id, product_name, description, category, language_code, base_unit_price, unit_of_measure, weight, dimensions, is_active, created_at, updated_at
-            FROM Products
-            WHERE {' AND '.join(sql_where_clauses)}
-            ORDER BY product_name
-        """
-
-        cursor.execute(sql, tuple(sql_params))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_products_for_selection: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-# Functions for ClientProjectProducts association
-def add_product_to_client_or_project(link_data: dict) -> int | None:
-    """Links a product to a client or project, calculating total price."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # Initialize effective_unit_price in case of early exit or error before its definition
-        effective_unit_price = None # Ensure it's in scope for the except block
-        product_info = None # Ensure it's in scope for the except block
-
-        product_id = link_data.get('product_id')
-        product_info = get_product_by_id(product_id) # Need this for base price
-        if not product_info:
-            print(f"Product with ID {product_id} not found.")
-            return None
-
-        quantity = link_data.get('quantity', 1)
-        unit_price_override = link_data.get('unit_price_override')
-
-        if unit_price_override is not None: # Explicit check for None
-            effective_unit_price = unit_price_override
-        else:
-            effective_unit_price = product_info.get('base_unit_price')
-
-        # Ensure effective_unit_price is not None before multiplication
-        if effective_unit_price is None:
-            # This case should ideally not be reached if base_unit_price is NOT NULL in DB
-            # and product_info is correctly fetched.
-            print(f"Warning: effective_unit_price was None for product ID {product_id} (Quantity: {quantity}, Override: {unit_price_override}, Base from DB: {product_info.get('base_unit_price') if product_info else 'N/A'}). Defaulting to 0.0.")
-            effective_unit_price = 0.0
-
-        total_price_calculated = quantity * effective_unit_price
-
-        sql = """
-            INSERT INTO ClientProjectProducts (
-                client_id, project_id, product_id, quantity, unit_price_override, total_price_calculated, added_at
-            ) VALUES (?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            link_data.get('client_id'),
-            link_data.get('project_id'), # Can be NULL
-            product_id,
-            quantity,
-            link_data.get('unit_price_override'), # Store override, or NULL if base used
-            total_price_calculated,
-            datetime.utcnow().isoformat() + "Z"
-        )
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.lastrowid
-    except TypeError as te: # Specifically catch the error we are investigating
-        print(f"TypeError in add_product_to_client_or_project: {te}")
-        print(f"  product_id: {link_data.get('product_id')}")
-        print(f"  quantity: {link_data.get('quantity', 1)}")
-        print(f"  unit_price_override from link_data: {link_data.get('unit_price_override')}")
-        if product_info: # product_info might be None if error happened before it was fetched
-            print(f"  base_unit_price from product_info: {product_info.get('base_unit_price')}")
-        else:
-            print(f"  product_info was None or not fetched prior to error.")
-        print(f"  effective_unit_price at point of error: {effective_unit_price if 'effective_unit_price' in locals() else 'Not yet defined or error before definition'}")
-        return None # Indicate failure
-    except sqlite3.Error as e:
-        print(f"Database error in add_product_to_client_or_project: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_products_for_client_or_project(client_id: str, project_id: str = None) -> list[dict]:
-    """Fetches products for a client, optionally filtered by project_id. Joins with Products."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        
-        sql = """
-            SELECT cpp.*,
-                   p.product_id as product_id_original_lang, p.product_name, p.description as product_description,
-                   p.category as product_category, p.base_unit_price, p.unit_of_measure,
-                   p.weight, p.dimensions, p.language_code
-            FROM ClientProjectProducts cpp
-            JOIN Products p ON cpp.product_id = p.product_id
-            WHERE cpp.client_id = ?
-        """
-        params = [client_id]
-        
-        if project_id:
-            sql += " AND cpp.project_id = ?"
-            params.append(project_id)
-        else: # Explicitly handle case where we want products not tied to any project for this client
-            sql += " AND cpp.project_id IS NULL"
-            
-        cursor.execute(sql, tuple(params))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_products_for_client_or_project: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def update_client_project_product(link_id: int, update_data: dict) -> bool:
-    """Updates a ClientProjectProduct link. Recalculates total_price if needed."""
-    conn = None
-    if not update_data: return False
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        # Fetch current link data to get product_id and existing values for price calculation
-        cursor.execute("SELECT * FROM ClientProjectProducts WHERE client_project_product_id = ?", (link_id,))
-        current_link = cursor.fetchone()
-        if not current_link:
-            print(f"ClientProjectProduct link with ID {link_id} not found.")
-            return False
-        
-        current_link_dict = dict(current_link)
-        new_quantity = update_data.get('quantity', current_link_dict['quantity'])
-        new_unit_price_override = update_data.get('unit_price_override', current_link_dict['unit_price_override'])
-
-        final_unit_price = new_unit_price_override
-        if final_unit_price is None: # If override is removed or was never there, use base price
-            product_info = get_product_by_id(current_link_dict['product_id'])
-            if not product_info: return False # Should not happen if data is consistent
-            final_unit_price = product_info['base_unit_price']
-        
-        update_data['total_price_calculated'] = new_quantity * final_unit_price
-        
-        set_clauses = [f"{key} = ?" for key in update_data.keys()]
-        params_list = list(update_data.values())
-        params_list.append(link_id)
-        
-        sql = f"UPDATE ClientProjectProducts SET {', '.join(set_clauses)} WHERE client_project_product_id = ?"
-        cursor.execute(sql, params_list)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in update_client_project_product: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-def remove_product_from_client_or_project(link_id: int) -> bool:
-    """Removes a product link from a client/project."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = "DELETE FROM ClientProjectProducts WHERE client_project_product_id = ?"
-        cursor.execute(sql, (link_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in remove_product_from_client_or_project: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
 # CRUD functions for ClientDocuments
 def add_client_document(doc_data: dict) -> str | None:
     """Adds a new client document. Returns document_id (UUID) or None."""
@@ -3565,193 +538,6 @@ def delete_client_document(document_id: str) -> bool:
         return False
     finally:
         if conn: conn.close()
-
-# CRUD functions for SmtpConfigs
-def _ensure_single_default_smtp(cursor: sqlite3.Cursor, exclude_id: int | None = None):
-    """Internal helper to ensure only one SMTP config is default."""
-    sql = "UPDATE SmtpConfigs SET is_default = FALSE WHERE is_default = TRUE"
-    if exclude_id is not None:
-        sql += " AND smtp_config_id != ?"
-        cursor.execute(sql, (exclude_id,))
-    else:
-        cursor.execute(sql)
-
-def add_smtp_config(config_data: dict) -> int | None:
-    """
-    Adds a new SMTP config. Returns smtp_config_id or None.
-    Expects 'password_encrypted'. Handles 'is_default' logic.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        # Use a transaction for default handling
-        conn.isolation_level = None # Explicitly start transaction for some Python versions / DB drivers
-        cursor = conn.cursor()
-        cursor.execute("BEGIN")
-
-        if config_data.get('is_default'):
-            _ensure_single_default_smtp(cursor)
-
-        now = datetime.utcnow().isoformat() + "Z" # Not in schema, but good practice if it were
-        sql = """
-            INSERT INTO SmtpConfigs (
-                config_name, smtp_server, smtp_port, username, password_encrypted,
-                use_tls, is_default, sender_email_address, sender_display_name
-            ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
-        """
-        params = (
-            config_data.get('config_name'), config_data.get('smtp_server'),
-            config_data.get('smtp_port'), config_data.get('username'),
-            config_data.get('password_encrypted'), # Assumed pre-encrypted
-            config_data.get('use_tls', True),
-            config_data.get('is_default', False),
-            config_data.get('sender_email_address'),
-            config_data.get('sender_display_name')
-        )
-        cursor.execute(sql, params)
-        new_id = cursor.lastrowid
-        conn.commit()
-        return new_id
-    except sqlite3.Error as e:
-        if conn: conn.rollback()
-        print(f"Database error in add_smtp_config: {e}")
-        return None
-    finally:
-        if conn: 
-            conn.isolation_level = '' # Reset to default
-            conn.close()
-
-
-def get_smtp_config_by_id(smtp_config_id: int) -> dict | None:
-    """Retrieves an SMTP config by ID."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM SmtpConfigs WHERE smtp_config_id = ?", (smtp_config_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_smtp_config_by_id: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_default_smtp_config() -> dict | None:
-    """Retrieves the default SMTP config."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM SmtpConfigs WHERE is_default = TRUE")
-        row = cursor.fetchone()
-        return dict(row) if row else None # Could be multiple if DB constraint not present, returns first
-    except sqlite3.Error as e:
-        print(f"Database error in get_default_smtp_config: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_all_smtp_configs() -> list[dict]:
-    """Retrieves all SMTP configs."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM SmtpConfigs ORDER BY config_name")
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_smtp_configs: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def update_smtp_config(smtp_config_id: int, config_data: dict) -> bool:
-    """
-    Updates an SMTP config. Handles 'is_default' logic.
-    Expects 'password_encrypted' if password is to be changed.
-    """
-    conn = None
-    if not config_data: return False
-    try:
-        conn = get_db_connection()
-        conn.isolation_level = None
-        cursor = conn.cursor()
-        cursor.execute("BEGIN")
-
-        if config_data.get('is_default'):
-            _ensure_single_default_smtp(cursor, exclude_id=smtp_config_id)
-        
-        # Ensure password_encrypted is handled if present, otherwise original remains
-        # This assumes if 'password_encrypted' is not in config_data, it's not being updated.
-        valid_columns = [
-            'config_name', 'smtp_server', 'smtp_port', 'username', 
-            'password_encrypted', 'use_tls', 'is_default', 
-            'sender_email_address', 'sender_display_name'
-        ]
-        current_config_data = {k: v for k,v in config_data.items() if k in valid_columns}
-        if not current_config_data:
-            conn.rollback() # No valid fields to update
-            return False
-
-        set_clauses = [f"{key} = ?" for key in current_config_data.keys()]
-        params = list(current_config_data.values())
-        params.append(smtp_config_id)
-        
-        sql = f"UPDATE SmtpConfigs SET {', '.join(set_clauses)} WHERE smtp_config_id = ?"
-        cursor.execute(sql, params)
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        if conn: conn.rollback()
-        print(f"Database error in update_smtp_config: {e}")
-        return False
-    finally:
-        if conn: 
-            conn.isolation_level = ''
-            conn.close()
-
-def delete_smtp_config(smtp_config_id: int) -> bool:
-    """Deletes an SMTP config."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # Consider logic if deleting the default config (e.g., pick another or ensure none is default)
-        # For now, direct delete.
-        cursor.execute("DELETE FROM SmtpConfigs WHERE smtp_config_id = ?", (smtp_config_id,))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"Database error in delete_smtp_config: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
-def set_default_smtp_config(smtp_config_id: int) -> bool:
-    """Sets a specific SMTP config as default and unsets others."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        conn.isolation_level = None
-        cursor = conn.cursor()
-        cursor.execute("BEGIN")
-        
-        _ensure_single_default_smtp(cursor, exclude_id=smtp_config_id)
-        cursor.execute("UPDATE SmtpConfigs SET is_default = TRUE WHERE smtp_config_id = ?", (smtp_config_id,))
-        updated_rows = cursor.rowcount
-        
-        conn.commit()
-        return updated_rows > 0
-    except sqlite3.Error as e:
-        if conn: conn.rollback()
-        print(f"Database error in set_default_smtp_config: {e}")
-        return False
-    finally:
-        if conn: 
-            conn.isolation_level = ''
-            conn.close()
 
 # --- ScheduledEmails Functions ---
 def add_scheduled_email(email_data: dict) -> int | None:
@@ -4165,37 +951,6 @@ def delete_kpi(kpi_id: int) -> bool:
         if conn:
             conn.close()
 
-# --- ApplicationSettings Functions ---
-def get_setting(key: str) -> str | None:
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT setting_value FROM ApplicationSettings WHERE setting_key = ?", (key,))
-        row = cursor.fetchone()
-        return row['setting_value'] if row else None
-    except sqlite3.Error as e:
-        print(f"DB error in get_setting: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def set_setting(key: str, value: str) -> bool:
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        # Using INSERT OR REPLACE (UPSERT)
-        sql = "INSERT OR REPLACE INTO ApplicationSettings (setting_key, setting_value) VALUES (?, ?)"
-        cursor.execute(sql, (key, value))
-        conn.commit()
-        return cursor.rowcount > 0
-    except sqlite3.Error as e:
-        print(f"DB error in set_setting: {e}")
-        return False
-    finally:
-        if conn: conn.close()
-
 # --- ActivityLog Functions ---
 def add_activity_log(log_data: dict) -> int | None:
     conn = None
@@ -4248,70 +1003,6 @@ def get_activity_logs(limit: int = 50, filters: dict = None) -> list[dict]:
         return []
     finally:
         if conn: conn.close()
-
-# --- Default Templates Population ---
-DEFAULT_COVER_PAGE_TEMPLATES = [
-    {
-        'template_name': 'Standard Report Cover',
-        'description': 'A standard cover page for general reports.',
-        'default_title': 'Report Title',
-        'default_subtitle': 'Company Subdivision',
-        'default_author': 'Automated Report Generator',
-        'style_config_json': {'font': 'Helvetica', 'primary_color': '#2a2a2a', 'secondary_color': '#5cb85c'},
-        'is_default_template': 1 # Mark as default
-    },
-    {
-        'template_name': 'Financial Statement Cover',
-        'description': 'Cover page for official financial statements.',
-        'default_title': 'Financial Statement',
-        'default_subtitle': 'Fiscal Year Ending YYYY',
-        'default_author': 'Finance Department',
-        'style_config_json': {'font': 'Times New Roman', 'primary_color': '#003366', 'secondary_color': '#e0a800'},
-        'is_default_template': 1 # Mark as default
-    },
-    {
-        'template_name': 'Creative Project Brief',
-        'description': 'A vibrant cover for creative project briefs and proposals.',
-        'default_title': 'Creative Brief: [Project Name]',
-        'default_subtitle': 'Client: [Client Name]',
-        'default_author': 'Creative Team',
-        'style_config_json': {'font': 'Montserrat', 'primary_color': '#ff6347', 'secondary_color': '#4682b4', 'layout_hint': 'two-column'},
-        'is_default_template': 1 # Mark as default
-    },
-    {
-        'template_name': 'Technical Document Cover',
-        'description': 'A clean and formal cover for technical documentation.',
-        'default_title': 'Technical Specification Document',
-        'default_subtitle': 'Version [VersionNumber]',
-        'default_author': 'Engineering Team',
-        'style_config_json': {'font': 'Roboto', 'primary_color': '#191970', 'secondary_color': '#cccccc'},
-        'is_default_template': 1 # Mark as default
-    }
-]
-
-def _populate_default_cover_page_templates():
-    """
-    Populates the CoverPageTemplates table with predefined default templates
-    if they do not already exist by name.
-    """
-    print("Attempting to populate default cover page templates...")
-    # Optionally, fetch a system user ID if you want to set created_by_user_id
-    # system_user = get_user_by_username('system_user') # Define or fetch a system user
-    # system_user_id = system_user['user_id'] if system_user else None
-
-    for template_def in DEFAULT_COVER_PAGE_TEMPLATES:
-        existing_template = get_cover_page_template_by_name(template_def['template_name'])
-        if existing_template:
-            print(f"Default template '{template_def['template_name']}' already exists. Skipping.")
-        else:
-            # template_def_with_user = {**template_def, 'created_by_user_id': system_user_id}
-            # new_id = add_cover_page_template(template_def_with_user)
-            new_id = add_cover_page_template(template_def) # Simpler: no user_id for defaults for now
-            if new_id:
-                print(f"Added default cover page template: '{template_def['template_name']}' with ID: {new_id}")
-            else:
-                print(f"Failed to add default cover page template: '{template_def['template_name']}'")
-    print("Default cover page templates population attempt finished.")
 
 # --- CoverPageTemplates CRUD ---
 def add_cover_page_template(template_data: dict) -> str | None:
@@ -4686,179 +1377,6 @@ def get_cover_pages_for_user(user_id: str, limit: int = 50, offset: int = 0) -> 
         if conn:
             conn.close()
 
-# --- Lookup Table GET Functions ---
-def get_all_countries() -> list[dict]:
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Countries ORDER BY country_name")
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"DB error in get_all_countries: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def get_country_by_id(country_id: int) -> dict | None:
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Countries WHERE country_id = ?", (country_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"DB error in get_country_by_id: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
-def get_country_by_name(country_name: str) -> dict | None:
-    """Retrieves a country by its name. Returns a dict or None if not found."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Countries WHERE country_name = ?", (country_name,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_country_by_name: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def add_country(country_data: dict) -> int | None:
-    """
-    Adds a new country to the Countries table.
-    Expects country_data to contain 'country_name'.
-    Returns the country_id of the newly added or existing country.
-    Returns None if an unexpected error occurs.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        country_name = country_data.get('country_name')
-        if not country_name:
-            print("Error: 'country_name' is required to add a country.")
-            return None
-
-        try:
-            cursor.execute("INSERT INTO Countries (country_name) VALUES (?)", (country_name,))
-            conn.commit()
-            return cursor.lastrowid
-        except sqlite3.IntegrityError:
-            # Country name already exists, fetch its ID
-            print(f"Country '{country_name}' already exists. Fetching its ID.")
-            cursor.execute("SELECT country_id FROM Countries WHERE country_name = ?", (country_name,))
-            row = cursor.fetchone()
-            return row['country_id'] if row else None
-
-    except sqlite3.Error as e:
-        print(f"Database error in add_country: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_cities(country_id: int = None) -> list[dict]:
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        sql = "SELECT * FROM Cities"
-        params = []
-        if country_id is not None:
-            sql += " WHERE country_id = ?"
-            params.append(country_id)
-        sql += " ORDER BY city_name"
-        cursor.execute(sql, params)
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"DB error in get_all_cities: {e}")
-        return []
-    finally:
-        if conn: conn.close()
-
-def add_city(city_data: dict) -> int | None:
-    """
-    Adds a new city to the Cities table.
-    Expects city_data to contain 'country_id' and 'city_name'.
-    Returns the city_id of the newly added or existing city.
-    Returns None if an unexpected error occurs or if country_id or city_name is missing.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        country_id = city_data.get('country_id')
-        city_name = city_data.get('city_name')
-
-        if not country_id or not city_name:
-            print("Error: 'country_id' and 'city_name' are required to add a city.")
-            return None
-
-        try:
-            # Check if city already exists for this country to avoid IntegrityError for composite uniqueness if not explicitly handled by schema (though schema doesn't show composite unique constraint for city_name+country_id, it's good practice)
-            # However, the current schema for Cities does not have a UNIQUE constraint on (country_id, city_name).
-            # It only has city_id as PK and country_id as FK.
-            # So, we will just insert. If a stricter uniqueness is needed, the table schema should be updated.
-            cursor.execute("INSERT INTO Cities (country_id, city_name) VALUES (?, ?)", (country_id, city_name))
-            conn.commit()
-            return cursor.lastrowid
-        except sqlite3.IntegrityError:
-            # This part would be relevant if there was a UNIQUE constraint on (country_id, city_name)
-            # For now, this block might not be hit unless city_name itself becomes unique across all countries (which is not typical)
-            print(f"IntegrityError likely means city '{city_name}' under country_id '{country_id}' already exists or other constraint failed.")
-            # If it were unique and we wanted to return existing:
-            # cursor.execute("SELECT city_id FROM Cities WHERE country_id = ? AND city_name = ?", (country_id, city_name))
-            # row = cursor.fetchone()
-            # return row['city_id'] if row else None
-            return None # For now, any IntegrityError is treated as a failure to add as new.
-
-    except sqlite3.Error as e:
-        print(f"Database error in add_city: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_city_by_name_and_country_id(city_name: str, country_id: int) -> dict | None:
-    """Retrieves a specific city by name for a given country_id."""
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Cities WHERE city_name = ? AND country_id = ?", (city_name, country_id))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"Database error in get_city_by_name_and_country_id: {e}")
-        return None
-    finally:
-        if conn:
-            conn.close()
-
-def get_city_by_id(city_id: int) -> dict | None:
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-        cursor.execute("SELECT * FROM Cities WHERE city_id = ?", (city_id,))
-        row = cursor.fetchone()
-        return dict(row) if row else None
-    except sqlite3.Error as e:
-        print(f"DB error in get_city_by_id: {e}")
-        return None
-    finally:
-        if conn: conn.close()
-
 def get_all_templates(template_type_filter: str = None, language_code_filter: str = None) -> list[dict]:
     """
     Retrieves all templates, optionally filtered by template_type and/or language_code.
@@ -4943,48 +1461,6 @@ def get_templates_by_category_id(category_id: int) -> list[dict]:
         return [dict(row) for row in rows]
     except sqlite3.Error as e:
         print(f"Database error in get_templates_by_category_id: {e}")
-        return []
-    finally:
-        if conn:
-            conn.close()
-
-def get_all_products_for_selection_filtered(language_code: str = None, name_pattern: str = None) -> list[dict]:
-    """
-    Retrieves active products for selection, optionally filtered by language_code
-    and/or name_pattern (searches product_name and description).
-    Orders by product_name.
-    """
-    conn = None
-    try:
-        conn = get_db_connection()
-        cursor = conn.cursor()
-
-        sql_params = []
-        # Start with base selection of active products
-        sql_where_clauses = ["is_active = TRUE"]
-
-        if language_code:
-            sql_where_clauses.append("language_code = ?")
-            sql_params.append(language_code)
-
-        if name_pattern:
-            # The name_pattern should be passed with wildcards already, e.g., "%search_term%"
-            sql_where_clauses.append("(product_name LIKE ? OR description LIKE ?)")
-            sql_params.append(name_pattern)
-            sql_params.append(name_pattern)
-
-        sql = f"""
-            SELECT product_id, product_name, description, base_unit_price, language_code
-            FROM Products
-            WHERE {' AND '.join(sql_where_clauses)}
-            ORDER BY product_name
-        """
-
-        cursor.execute(sql, tuple(sql_params))
-        rows = cursor.fetchall()
-        return [dict(row) for row in rows]
-    except sqlite3.Error as e:
-        print(f"Database error in get_all_products_for_selection_filtered: {e}")
         return []
     finally:
         if conn:
